Network.AWS.Glacier

Description

Amazon Glacier is a secure, durable, and extremely low-cost storage service for data archiving and online backup. Customers can reliably store large or small amounts of data for as little as $0.01 per gigabyte per month, a significant savings compared to on-premises solutions. To keep costs low, Amazon Glacier is optimized for infrequently accessed data where a retrieval time of several hours is suitable.

Documentation

module Network.AWS.Glacier.AbortMultipartUpload

module Network.AWS.Glacier.CompleteMultipartUpload

module Network.AWS.Glacier.CreateVault

module Network.AWS.Glacier.DeleteArchive

module Network.AWS.Glacier.DeleteVault

module Network.AWS.Glacier.DeleteVaultNotifications

module Network.AWS.Glacier.DescribeJob

module Network.AWS.Glacier.DescribeVault

module Network.AWS.Glacier.GetDataRetrievalPolicy

module Network.AWS.Glacier.GetJobOutput

module Network.AWS.Glacier.GetVaultNotifications

module Network.AWS.Glacier.InitiateJob

module Network.AWS.Glacier.InitiateMultipartUpload

module Network.AWS.Glacier.ListJobs

module Network.AWS.Glacier.ListMultipartUploads

module Network.AWS.Glacier.ListParts

module Network.AWS.Glacier.ListVaults

module Network.AWS.Glacier.SetDataRetrievalPolicy

module Network.AWS.Glacier.SetVaultNotifications

module Network.AWS.Glacier.Types

module Network.AWS.Glacier.UploadArchive

module Network.AWS.Glacier.UploadMultipartPart