Network.AWS.Glacier.AbortMultipartUpload

Description

This operation aborts a multipart upload identified by the upload ID.

After the Abort Multipart Upload request succeeds, you cannot upload any more parts to the multipart upload or complete the multipart upload. Aborting a completed upload fails. However, aborting an already-aborted upload will succeed, for a short time. For more information about uploading a part and completing a multipart upload, see UploadMultipartPart and CompleteMultipartUpload.

This operation is idempotent.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identityand Access Management (IAM).

For conceptual information and underlying REST API, go to Working withArchives in Amazon Glacier and Abort Multipart Upload in the Amazon GlacierDeveloper Guide.

http://docs.aws.amazon.com/amazonglacier/latest/dev/api-AbortMultipartUpload.html

Request

data AbortMultipartUpload

Request constructor

abortMultipartUpload

Arguments

:: Text
amuAccountId

-> Text
amuVaultName

-> Text
amuUploadId

-> AbortMultipartUpload

AbortMultipartUpload constructor.

The fields accessible through corresponding lenses are:

amuAccountId :: Text

amuUploadId :: Text

amuVaultName :: Text

Request lenses

amuAccountId :: Lens' AbortMultipartUpload Text

The AccountId is the AWS Account ID. You can specify either the AWS Account ID or optionally a -, in which case Amazon Glacier uses the AWS Account ID associated with the credentials used to sign the request. If you specify your Account ID, do not include hyphens in it.

amuUploadId :: Lens' AbortMultipartUpload Text

The upload ID of the multipart upload to delete.

amuVaultName :: Lens' AbortMultipartUpload Text

The name of the vault.

Response

data AbortMultipartUploadResponse

Response constructor

abortMultipartUploadResponse :: AbortMultipartUploadResponse

AbortMultipartUploadResponse constructor.