Network.AWS.Glacier.AddTagsToVault

Description

This operation adds the specified tags to a vault. Each tag is composed of a key and a value. Each vault can have up to 10 tags. If your request would cause the tag limit for the vault to be exceeded, the operation throws the LimitExceededException error. If a tag already exists on the vault under a specified key, the existing key value will be overwritten. For more information about tags, see Tagging Amazon Glacier Resources .

Creating a Request

addTagsToVault

Arguments

:: Text
attvAccountId

-> Text
attvVaultName

-> AddTagsToVault

Creates a value of AddTagsToVault with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

attvTags - The tags to add to the vault. Each tag is composed of a key and a value. The value can be an empty string.

attvAccountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

attvVaultName - The name of the vault.

data AddTagsToVault

The input values for AddTagsToVault .

See: addTagsToVault smart constructor.

Request Lenses

attvTags :: Lens' AddTagsToVault (HashMap Text Text)

The tags to add to the vault. Each tag is composed of a key and a value. The value can be an empty string.

attvAccountId :: Lens' AddTagsToVault Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

attvVaultName :: Lens' AddTagsToVault Text

The name of the vault.

Destructuring the Response

addTagsToVaultResponse :: AddTagsToVaultResponse

Creates a value of AddTagsToVaultResponse with the minimum fields required to make a request.

data AddTagsToVaultResponse

See: addTagsToVaultResponse smart constructor.