Network.AWS.Glacier.CompleteMultipartUpload

Description

You call this operation to inform Amazon Glacier that all the archive parts have been uploaded and that Amazon Glacier can now assemble the archive from the uploaded parts. After assembling and saving the archive to the vault, Amazon Glacier returns the URI path of the newly created archive resource. Using the URI path, you can then access the archive. After you upload an archive, you should save the archive ID returned to retrieve the archive at a later point. You can also get the vault inventory to obtain a list of archive IDs in a vault. For more information, see InitiateJob.

In the request, you must include the computed SHA256 tree hash of the entire archive you have uploaded. For information about computing a SHA256 tree hash, see Computing Checksums. On the server side, Amazon Glacier also constructs the SHA256 tree hash of the assembled archive. If the values match, Amazon Glacier saves the archive to the vault; otherwise, it returns an error, and the operation fails. The ListParts operation returns a list of parts uploaded for a specific multipart upload. It includes checksum information for each uploaded part that can be used to debug a bad checksum issue.

Additionally, Amazon Glacier also checks for any missing content ranges when assembling the archive, if missing content ranges are found, Amazon Glacier returns an error and the operation fails.

Complete Multipart Upload is an idempotent operation. After your first successful complete multipart upload, if you call the operation again within a short period, the operation will succeed and return the same archive ID. This is useful in the event you experience a network issue that causes an aborted connection or receive a 500 server error, in which case you can repeat your Complete Multipart Upload request and get the same archive ID without creating duplicate archives. Note, however, that after the multipart upload completes, you cannot call the List Parts operation and the multipart upload will not appear in List Multipart Uploads response, even if idempotent complete is possible.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identityand Access Management (IAM).

For conceptual information and underlying REST API, go to Uploading LargeArchives in Parts (Multipart Upload) and Complete Multipart Upload in the Amazon Glacier Developer Guide.

http://docs.aws.amazon.com/amazonglacier/latest/dev/api-CompleteMultipartUpload.html

Request

data CompleteMultipartUpload

Request constructor

completeMultipartUpload

Arguments

:: Text
cmuAccountId

-> Text
cmuVaultName

-> Text
cmuUploadId

-> CompleteMultipartUpload

CompleteMultipartUpload constructor.

The fields accessible through corresponding lenses are:

cmuAccountId :: Text

cmuArchiveSize :: Maybe Text

cmuChecksum :: Maybe Text

cmuUploadId :: Text

cmuVaultName :: Text

Request lenses

cmuAccountId :: Lens' CompleteMultipartUpload Text

The AccountId is the AWS Account ID. You can specify either the AWS Account ID or optionally a -, in which case Amazon Glacier uses the AWS Account ID associated with the credentials used to sign the request. If you specify your Account ID, do not include hyphens in it.

cmuArchiveSize :: Lens' CompleteMultipartUpload (Maybe Text)

The total size, in bytes, of the entire archive. This value should be the sum of all the sizes of the individual parts that you uploaded.

cmuChecksum :: Lens' CompleteMultipartUpload (Maybe Text)

The SHA256 tree hash of the entire archive. It is the tree hash of SHA256 tree hash of the individual parts. If the value you specify in the request does not match the SHA256 tree hash of the final assembled archive as computed by Amazon Glacier, Amazon Glacier returns an error and the request fails.

cmuUploadId :: Lens' CompleteMultipartUpload Text

The upload ID of the multipart upload.

cmuVaultName :: Lens' CompleteMultipartUpload Text

The name of the vault.

Response

data CompleteMultipartUploadResponse

Response constructor

completeMultipartUploadResponse :: CompleteMultipartUploadResponse

CompleteMultipartUploadResponse constructor.

The fields accessible through corresponding lenses are:

cmurArchiveId :: Maybe Text

cmurChecksum :: Maybe Text

cmurLocation :: Maybe Text

Response lenses

cmurArchiveId :: Lens' CompleteMultipartUploadResponse (Maybe Text)

The ID of the archive. This value is also included as part of the location.

cmurChecksum :: Lens' CompleteMultipartUploadResponse (Maybe Text)

The checksum of the archive computed by Amazon Glacier.

cmurLocation :: Lens' CompleteMultipartUploadResponse (Maybe Text)

The relative URI path of the newly added archive resource.