Network.AWS.Glacier.CompleteVaultLock

Description

This operation completes the vault locking process by transitioning the vault lock from the InProgress state to the Locked state, which causes the vault lock policy to become unchangeable. A vault lock is put into the InProgress state by calling InitiateVaultLock . You can obtain the state of the vault lock by calling GetVaultLock . For more information about the vault locking process, Amazon Glacier Vault Lock .

This operation is idempotent. This request is always successful if the vault lock is in the Locked state and the provided lock ID matches the lock ID originally used to lock the vault.

If an invalid lock ID is passed in the request when the vault lock is in the Locked state, the operation returns an AccessDeniedException error. If an invalid lock ID is passed in the request when the vault lock is in the InProgress state, the operation throws an InvalidParameter error.

Creating a Request

completeVaultLock

Arguments

:: Text
cvlAccountId

-> Text
cvlVaultName

-> Text
cvlLockId

-> CompleteVaultLock

Creates a value of CompleteVaultLock with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cvlAccountId - The AccountId value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens (-) in the ID.

cvlVaultName - The name of the vault.

cvlLockId - The lockId value is the lock ID obtained from a InitiateVaultLock request.

data CompleteVaultLock

The input values for CompleteVaultLock .

See: completeVaultLock smart constructor.

Request Lenses

cvlAccountId :: Lens' CompleteVaultLock Text

The AccountId value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens (-) in the ID.

cvlVaultName :: Lens' CompleteVaultLock Text

The name of the vault.

cvlLockId :: Lens' CompleteVaultLock Text

The lockId value is the lock ID obtained from a InitiateVaultLock request.

Destructuring the Response

completeVaultLockResponse :: CompleteVaultLockResponse

Creates a value of CompleteVaultLockResponse with the minimum fields required to make a request.

data CompleteVaultLockResponse

See: completeVaultLockResponse smart constructor.