Network.AWS.Glacier.CreateVault

Description

This operation creates a new vault with the specified name. The name of the vault must be unique within a region for an AWS account. You can create up to 1,000 vaults per account. If you need to create more vaults, contact Amazon Glacier.

You must use the following guidelines when naming a vault.

Names can be between 1 and 255 characters long.

Allowed characters are a-z, A-Z, 0-9, '_' (underscore), - (hyphen), and . (period).

This operation is idempotent.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identityand Access Management (IAM).

For conceptual information and underlying REST API, go to Creating a Vaultin Amazon Glacier and Create Vault in the Amazon Glacier Developer Guide.

http://docs.aws.amazon.com/amazonglacier/latest/dev/api-CreateVault.html

Request

data CreateVault

Request constructor

createVault

Arguments

:: Text
cvAccountId

-> Text
cvVaultName

-> CreateVault

CreateVault constructor.

The fields accessible through corresponding lenses are:

cvAccountId :: Text

cvVaultName :: Text

Request lenses

cvAccountId :: Lens' CreateVault Text

The AccountId is the AWS Account ID. You can specify either the AWS Account ID or optionally a -, in which case Amazon Glacier uses the AWS Account ID associated with the credentials used to sign the request. If you specify your Account ID, do not include hyphens in it.

cvVaultName :: Lens' CreateVault Text

The name of the vault.

Response

data CreateVaultResponse

Response constructor

createVaultResponse :: CreateVaultResponse

CreateVaultResponse constructor.

The fields accessible through corresponding lenses are:

cvrLocation :: Maybe Text

Response lenses

cvrLocation :: Lens' CreateVaultResponse (Maybe Text)

The URI of the vault that was created.