Network.AWS.Glacier.DeleteArchive

Description

This operation deletes an archive from a vault. Subsequent requests to initiate a retrieval of this archive will fail. Archive retrievals that are in progress for this archive ID may or may not succeed according to the following scenarios:

If the archive retrieval job is actively preparing the data for download when Amazon Glacier receives the delete archive request, the archival retrieval operation might fail. If the archive retrieval job has successfully prepared the archive for download when Amazon Glacier receives the delete archive request, you will be able to download the output. This operation is idempotent. Attempting to delete an already-deleted archive does not result in an error.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identityand Access Management (IAM).

For conceptual information and underlying REST API, go to Deleting anArchive in Amazon Glacier and Delete Archive in the Amazon Glacier DeveloperGuide.

http://docs.aws.amazon.com/amazonglacier/latest/dev/api-DeleteArchive.html

Request

data DeleteArchive

Request constructor

deleteArchive

Arguments

:: Text
daAccountId

-> Text
daVaultName

-> Text
daArchiveId

-> DeleteArchive

DeleteArchive constructor.

The fields accessible through corresponding lenses are:

daAccountId :: Text

daArchiveId :: Text

daVaultName :: Text

Request lenses

daAccountId :: Lens' DeleteArchive Text

The AccountId is the AWS Account ID. You can specify either the AWS Account ID or optionally a -, in which case Amazon Glacier uses the AWS Account ID associated with the credentials used to sign the request. If you specify your Account ID, do not include hyphens in it.

daArchiveId :: Lens' DeleteArchive Text

The ID of the archive to delete.

daVaultName :: Lens' DeleteArchive Text

The name of the vault.

Response

data DeleteArchiveResponse

Response constructor

deleteArchiveResponse :: DeleteArchiveResponse

DeleteArchiveResponse constructor.