Network.AWS.Glacier.DeleteVault

Description

This operation deletes a vault. Amazon Glacier will delete a vault only if there are no archives in the vault as of the last inventory and there have been no writes to the vault since the last inventory. If either of these conditions is not satisfied, the vault deletion fails (that is, the vault is not removed) and Amazon Glacier returns an error. You can use DescribeVault to return the number of archives in a vault, and you can use Initiate a Job(POST jobs) to initiate a new inventory retrieval for a vault. The inventory contains the archive IDs you use to delete archives using Delete Archive(DELETE archive).

This operation is idempotent.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identityand Access Management (IAM).

For conceptual information and underlying REST API, go to Deleting a Vaultin Amazon Glacier and Delete Vault in the Amazon Glacier Developer Guide.

http://docs.aws.amazon.com/amazonglacier/latest/dev/api-DeleteVault.html

Request

data DeleteVault

Request constructor

deleteVault

Arguments

:: Text
dv1AccountId

-> Text
dv1VaultName

-> DeleteVault

DeleteVault constructor.

The fields accessible through corresponding lenses are:

dv1AccountId :: Text

dv1VaultName :: Text

Request lenses

dv1AccountId :: Lens' DeleteVault Text

The AccountId is the AWS Account ID. You can specify either the AWS Account ID or optionally a -, in which case Amazon Glacier uses the AWS Account ID associated with the credentials used to sign the request. If you specify your Account ID, do not include hyphens in it.

dv1VaultName :: Lens' DeleteVault Text

The name of the vault.

Response

data DeleteVaultResponse

Response constructor

deleteVaultResponse :: DeleteVaultResponse

DeleteVaultResponse constructor.