Network.AWS.Glacier.DeleteVaultNotifications

Description

This operation deletes the notification configuration set for a vault. The operation is eventually consistent;that is, it might take some time for Amazon Glacier to completely disable the notifications and you might still receive some notifications for a short time after you send the delete request.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identityand Access Management (IAM).

For conceptual information and underlying REST API, go to Configuring VaultNotifications in Amazon Glacier and Delete Vault Notification Configuration in the Amazon Glacier Developer Guide.

http://docs.aws.amazon.com/amazonglacier/latest/dev/api-DeleteVaultNotifications.html

Request

data DeleteVaultNotifications

Request constructor

deleteVaultNotifications

Arguments

:: Text
dvnAccountId

-> Text
dvnVaultName

-> DeleteVaultNotifications

DeleteVaultNotifications constructor.

The fields accessible through corresponding lenses are:

dvnAccountId :: Text

dvnVaultName :: Text

Request lenses

dvnAccountId :: Lens' DeleteVaultNotifications Text

The AccountId is the AWS Account ID. You can specify either the AWS Account ID or optionally a -, in which case Amazon Glacier uses the AWS Account ID associated with the credentials used to sign the request. If you specify your Account ID, do not include hyphens in it.

dvnVaultName :: Lens' DeleteVaultNotifications Text

The name of the vault.

Response

data DeleteVaultNotificationsResponse

Response constructor

deleteVaultNotificationsResponse :: DeleteVaultNotificationsResponse

DeleteVaultNotificationsResponse constructor.