Network.AWS.Glacier.DescribeJob

Description

This operation returns information about a job you previously initiated, including the job initiation date, the user who initiated the job, the job status code/message and the Amazon SNS topic to notify after Amazon Glacier completes the job. For more information about initiating a job, see InitiateJob.

This operation enables you to check the status of your job. However, it is strongly recommended that you set up an Amazon SNS topic and specify it in your initiate job request so that Amazon Glacier can notify the topic after it completes the job.

A job ID will not expire for at least 24 hours after Amazon Glacier completes the job.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identityand Access Management (IAM).

For information about the underlying REST API, go to Working with Archivesin Amazon Glacier in the Amazon Glacier Developer Guide.

http://docs.aws.amazon.com/amazonglacier/latest/dev/api-DescribeJob.html

Request

data DescribeJob

Request constructor

describeJob

Arguments

:: Text
djAccountId

-> Text
djVaultName

-> Text
djJobId

-> DescribeJob

DescribeJob constructor.

The fields accessible through corresponding lenses are:

djAccountId :: Text

djJobId :: Text

djVaultName :: Text

Request lenses

djAccountId :: Lens' DescribeJob Text

The AccountId is the AWS Account ID. You can specify either the AWS Account ID or optionally a -, in which case Amazon Glacier uses the AWS Account ID associated with the credentials used to sign the request. If you specify your Account ID, do not include hyphens in it.

djJobId :: Lens' DescribeJob Text

The ID of the job to describe.

djVaultName :: Lens' DescribeJob Text

The name of the vault.

Response

data DescribeJobResponse

Response constructor

describeJobResponse :: DescribeJobResponse

DescribeJobResponse constructor.

The fields accessible through corresponding lenses are:

djrAction :: Maybe ActionCode

djrArchiveId :: Maybe Text

djrArchiveSHA256TreeHash :: Maybe Text

djrArchiveSizeInBytes :: Maybe Integer

djrCompleted :: Maybe Bool

djrCompletionDate :: Maybe Text

djrCreationDate :: Maybe Text

djrInventoryRetrievalParameters :: Maybe InventoryRetrievalJobDescription

djrInventorySizeInBytes :: Maybe Integer

djrJobDescription :: Maybe Text

djrJobId :: Maybe Text

djrRetrievalByteRange :: Maybe Text

djrSHA256TreeHash :: Maybe Text

djrSNSTopic :: Maybe Text

djrStatusCode :: Maybe StatusCode

djrStatusMessage :: Maybe Text

djrVaultARN :: Maybe Text

Response lenses

djrAction :: Lens' DescribeJobResponse (Maybe ActionCode)

The job type. It is either ArchiveRetrieval or InventoryRetrieval.

djrArchiveId :: Lens' DescribeJobResponse (Maybe Text)

For an ArchiveRetrieval job, this is the archive ID requested for download. Otherwise, this field is null.

djrArchiveSHA256TreeHash :: Lens' DescribeJobResponse (Maybe Text)

The SHA256 tree hash of the entire archive for an archive retrieval. For inventory retrieval jobs, this field is null.

djrArchiveSizeInBytes :: Lens' DescribeJobResponse (Maybe Integer)

For an ArchiveRetrieval job, this is the size in bytes of the archive being requested for download. For the InventoryRetrieval job, the value is null.

djrCompleted :: Lens' DescribeJobResponse (Maybe Bool)

The job status. When a job is completed, you get the job's output.

djrCompletionDate :: Lens' DescribeJobResponse (Maybe Text)

The UTC time that the archive retrieval request completed. While the job is in progress, the value will be null.

djrCreationDate :: Lens' DescribeJobResponse (Maybe Text)

The UTC date when the job was created. A string representation of ISO 8601 date format, for example, "2012-03-20T17:03:43.221Z".

djrInventoryRetrievalParameters :: Lens' DescribeJobResponse (Maybe InventoryRetrievalJobDescription)

Parameters used for range inventory retrieval.

djrInventorySizeInBytes :: Lens' DescribeJobResponse (Maybe Integer)

For an InventoryRetrieval job, this is the size in bytes of the inventory requested for download. For the ArchiveRetrieval job, the value is null.

djrJobDescription :: Lens' DescribeJobResponse (Maybe Text)

The job description you provided when you initiated the job.

djrJobId :: Lens' DescribeJobResponse (Maybe Text)

An opaque string that identifies an Amazon Glacier job.

djrRetrievalByteRange :: Lens' DescribeJobResponse (Maybe Text)

The retrieved byte range for archive retrieval jobs in the form "StartByteValue-EndByteValue" If no range was specified in the archive retrieval, then the whole archive is retrieved and StartByteValue equals 0 and EndByteValue equals the size of the archive minus 1. For inventory retrieval jobs this field is null.

djrSHA256TreeHash :: Lens' DescribeJobResponse (Maybe Text)

For an ArchiveRetrieval job, it is the checksum of the archive. Otherwise, the value is null.

The SHA256 tree hash value for the requested range of an archive. If the Initiate a Job request for an archive specified a tree-hash aligned range, then this field returns a value.

For the specific case when the whole archive is retrieved, this value is the same as the ArchiveSHA256TreeHash value.

This field is null in the following situations: Archive retrieval jobs that specify a range that is not tree-hash aligned.

Archival jobs that specify a range that is equal to the whole archive and the job status is InProgress.

Inventory jobs.

djrSNSTopic :: Lens' DescribeJobResponse (Maybe Text)

An Amazon Simple Notification Service (Amazon SNS) topic that receives notification.

djrStatusCode :: Lens' DescribeJobResponse (Maybe StatusCode)

The status code can be InProgress, Succeeded, or Failed, and indicates the status of the job.

djrStatusMessage :: Lens' DescribeJobResponse (Maybe Text)

A friendly message that describes the job status.

djrVaultARN :: Lens' DescribeJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the vault from which the archive retrieval was requested.