Network.AWS.Glacier.DescribeVault

Description

This operation returns information about a vault, including the vault's Amazon Resource Name (ARN), the date the vault was created, the number of archives it contains, and the total size of all the archives in the vault. The number of archives and their total size are as of the last inventory generation. This means that if you add or remove an archive from a vault, and then immediately use Describe Vault, the change in contents will not be immediately reflected. If you want to retrieve the latest inventory of the vault, use InitiateJob. Amazon Glacier generates vault inventories approximately daily. For more information, see Downloading a Vault Inventoryin Amazon Glacier.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identityand Access Management (IAM).

For conceptual information and underlying REST API, go to Retrieving VaultMetadata in Amazon Glacier and Describe Vault in the Amazon GlacierDeveloper Guide.

http://docs.aws.amazon.com/amazonglacier/latest/dev/api-DescribeVault.html

Request

data DescribeVault

Request constructor

describeVault

Arguments

:: Text
dvAccountId

-> Text
dvVaultName

-> DescribeVault

DescribeVault constructor.

The fields accessible through corresponding lenses are:

dvAccountId :: Text

dvVaultName :: Text

Request lenses

dvAccountId :: Lens' DescribeVault Text

The AccountId is the AWS Account ID. You can specify either the AWS Account ID or optionally a -, in which case Amazon Glacier uses the AWS Account ID associated with the credentials used to sign the request. If you specify your Account ID, do not include hyphens in it.

dvVaultName :: Lens' DescribeVault Text

The name of the vault.

Response

data DescribeVaultResponse

Response constructor

describeVaultResponse :: DescribeVaultResponse

DescribeVaultResponse constructor.

The fields accessible through corresponding lenses are:

dvrCreationDate :: Maybe Text

dvrLastInventoryDate :: Maybe Text

dvrNumberOfArchives :: Maybe Integer

dvrSizeInBytes :: Maybe Integer

dvrVaultARN :: Maybe Text

dvrVaultName :: Maybe Text

Response lenses

dvrCreationDate :: Lens' DescribeVaultResponse (Maybe Text)

The UTC date when the vault was created. A string representation of ISO 8601 date format, for example, "2012-03-20T17:03:43.221Z".

dvrLastInventoryDate :: Lens' DescribeVaultResponse (Maybe Text)

The UTC date when Amazon Glacier completed the last vault inventory. A string representation of ISO 8601 date format, for example, "2012-03-20T17:03:43.221Z".

dvrNumberOfArchives :: Lens' DescribeVaultResponse (Maybe Integer)

The number of archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example, if you just created the vault.

dvrSizeInBytes :: Lens' DescribeVaultResponse (Maybe Integer)

Total size, in bytes, of the archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example, if you just created the vault.

dvrVaultARN :: Lens' DescribeVaultResponse (Maybe Text)

The Amazon Resource Name (ARN) of the vault.

dvrVaultName :: Lens' DescribeVaultResponse (Maybe Text)

The name of the vault.