Network.AWS.Glacier.GetDataRetrievalPolicy

Description

This operation returns the current data retrieval policy for the account and region specified in the GET request. For more information about data retrieval policies, see Amazon Glacier Data Retrieval Policies .

Creating a Request

getDataRetrievalPolicy

Arguments

:: Text
gdrpAccountId

-> GetDataRetrievalPolicy

Creates a value of GetDataRetrievalPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdrpAccountId - The AccountId value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens (-) in the ID.

data GetDataRetrievalPolicy

Input for GetDataRetrievalPolicy.

See: getDataRetrievalPolicy smart constructor.

Request Lenses

gdrpAccountId :: Lens' GetDataRetrievalPolicy Text

The AccountId value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens (-) in the ID.

Destructuring the Response

getDataRetrievalPolicyResponse

Arguments

:: Int
gdrprsResponseStatus

-> GetDataRetrievalPolicyResponse

Creates a value of GetDataRetrievalPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdrprsPolicy - Contains the returned data retrieval policy in JSON format.

gdrprsResponseStatus - -- | The response status code.

data GetDataRetrievalPolicyResponse

Contains the Amazon Glacier response to the GetDataRetrievalPolicy request.

See: getDataRetrievalPolicyResponse smart constructor.

Response Lenses

gdrprsPolicy :: Lens' GetDataRetrievalPolicyResponse (Maybe DataRetrievalPolicy)

Contains the returned data retrieval policy in JSON format.

gdrprsResponseStatus :: Lens' GetDataRetrievalPolicyResponse Int

- | The response status code.