Network.AWS.Glacier.GetJobOutput

Description

This operation downloads the output of the job you initiated using InitiateJob . Depending on the job type you specified when you initiated the job, the output will be either the content of an archive or a vault inventory.

A job ID will not expire for at least 24 hours after Amazon Glacier completes the job. That is, you can download the job output within the 24 hours period after Amazon Glacier completes the job.

If the job output is large, then you can use the Range request header to retrieve a portion of the output. This allows you to download the entire output in smaller chunks of bytes. For example, suppose you have 1 GB of job output you want to download and you decide to download 128 MB chunks of data at a time, which is a total of eight Get Job Output requests. You use the following process to download the job output:

Download a 128 MB chunk of output by specifying the appropriate byte range using the Range header.

Along with the data, the response includes a checksum of the payload. You compute the checksum of the payload on the client and compare it with the checksum you received in the response to ensure you received all the expected data.

Repeat steps 1 and 2 for all the eight 128 MB chunks of output data, each time specifying the appropriate byte range.

After downloading all the parts of the job output, you have a list of eight checksum values. Compute the tree hash of these values to find the checksum of the entire output. Using the DescribeJob API, obtain job information of the job that provided you the output. The response includes the checksum of the entire archive stored in Amazon Glacier. You compare this value with the checksum you computed to ensure you have downloaded the entire archive content with no errors.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identityand Access Management (IAM).

For conceptual information and the underlying REST API, go to Downloading aVault Inventory, Downloading an Archive, and Get Job Output

http://docs.aws.amazon.com/amazonglacier/latest/dev/api-GetJobOutput.html

Request

data GetJobOutput

Request constructor

getJobOutput

Arguments

:: Text
gjoAccountId

-> Text
gjoVaultName

-> Text
gjoJobId

-> GetJobOutput

GetJobOutput constructor.

The fields accessible through corresponding lenses are:

gjoAccountId :: Text

gjoJobId :: Text

gjoRange :: Maybe Text

gjoVaultName :: Text

Request lenses

gjoAccountId :: Lens' GetJobOutput Text

The AccountId is the AWS Account ID. You can specify either the AWS Account ID or optionally a -, in which case Amazon Glacier uses the AWS Account ID associated with the credentials used to sign the request. If you specify your Account ID, do not include hyphens in it.

gjoJobId :: Lens' GetJobOutput Text

The job ID whose data is downloaded.

gjoRange :: Lens' GetJobOutput (Maybe Text)

The range of bytes to retrieve from the output. For example, if you want to download the first 1,048,576 bytes, specify "Range: bytes=0-1048575". By default, this operation downloads the entire output.

gjoVaultName :: Lens' GetJobOutput Text

The name of the vault.

Response

data GetJobOutputResponse

Response constructor

getJobOutputResponse

Arguments

:: RsBody
gjorBody

-> GetJobOutputResponse

GetJobOutputResponse constructor.

The fields accessible through corresponding lenses are:

gjorAcceptRanges :: Maybe Text

gjorArchiveDescription :: Maybe Text

gjorBody :: RsBody

gjorChecksum :: Maybe Text

gjorContentRange :: Maybe Text

gjorContentType :: Maybe Text

gjorStatus :: Maybe Int

Response lenses

gjorAcceptRanges :: Lens' GetJobOutputResponse (Maybe Text)

Indicates the range units accepted. For more information, go to RFC2616.

gjorArchiveDescription :: Lens' GetJobOutputResponse (Maybe Text)

The description of an archive.

gjorBody :: Lens' GetJobOutputResponse RsBody

The job data, either archive data or inventory data.

gjorChecksum :: Lens' GetJobOutputResponse (Maybe Text)

The checksum of the data in the response. This header is returned only when retrieving the output for an archive retrieval job. Furthermore, this header appears only under the following conditions: You get the entire range of the archive. You request a range to return of the archive that starts and ends on a multiple of 1 MB. For example, if you have an 3.1 MB archive and you specify a range to return that starts at 1 MB and ends at 2 MB, then the x-amz-sha256-tree-hash is returned as a response header. You request a range of the archive to return that starts on a multiple of 1 MB and goes to the end of the archive. For example, if you have a 3.1 MB archive and you specify a range that starts at 2 MB and ends at 3.1 MB (the end of the archive), then the x-amz-sha256-tree-hash is returned as a response header.

gjorContentRange :: Lens' GetJobOutputResponse (Maybe Text)

The range of bytes returned by Amazon Glacier. If only partial output is downloaded, the response provides the range of bytes Amazon Glacier returned. For example, bytes 0-1048575/8388608 returns the first 1 MB from 8 MB.

gjorContentType :: Lens' GetJobOutputResponse (Maybe Text)

The Content-Type depends on whether the job output is an archive or a vault inventory. For archive data, the Content-Type is application/octet-stream. For vault inventory, if you requested CSV format when you initiated the job, the Content-Type is text/csv. Otherwise, by default, vault inventory is returned as JSON, and the Content-Type is application/json.

gjorStatus :: Lens' GetJobOutputResponse (Maybe Int)

The HTTP response code for a job output request. The value depends on whether a range was specified in the request.