Network.AWS.Glacier.GetVaultNotifications

Description

This operation retrieves the 'notification-configuration' subresource of the specified vault.

For information about setting a notification configuration on a vault, see SetVaultNotifications. If a notification configuration for a vault is not set, the operation returns a '404 Not Found' error. For more information about vault notifications, see Configuring Vault Notifications in Amazon Glacier.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identityand Access Management (IAM).

For conceptual information and underlying REST API, go to Configuring VaultNotifications in Amazon Glacier and Get Vault Notification Configuration in the Amazon Glacier Developer Guide.

http://docs.aws.amazon.com/amazonglacier/latest/dev/api-GetVaultNotifications.html

Request

data GetVaultNotifications

Request constructor

getVaultNotifications

Arguments

:: Text
gvnAccountId

-> Text
gvnVaultName

-> GetVaultNotifications

GetVaultNotifications constructor.

The fields accessible through corresponding lenses are:

gvnAccountId :: Text

gvnVaultName :: Text

Request lenses

gvnAccountId :: Lens' GetVaultNotifications Text

The AccountId is the AWS Account ID. You can specify either the AWS Account ID or optionally a -, in which case Amazon Glacier uses the AWS Account ID associated with the credentials used to sign the request. If you specify your Account ID, do not include hyphens in it.

gvnVaultName :: Lens' GetVaultNotifications Text

The name of the vault.

Response

data GetVaultNotificationsResponse

Response constructor

getVaultNotificationsResponse :: GetVaultNotificationsResponse

GetVaultNotificationsResponse constructor.

The fields accessible through corresponding lenses are:

gvnrVaultNotificationConfig :: Maybe VaultNotificationConfig

Response lenses

gvnrVaultNotificationConfig :: Lens' GetVaultNotificationsResponse (Maybe VaultNotificationConfig)

Returns the notification configuration set on the vault.