Network.AWS.Glacier.InitiateMultipartUpload

Description

This operation initiates a multipart upload. Amazon Glacier creates a multipart upload resource and returns its ID in the response. The multipart upload ID is used in subsequent requests to upload parts of an archive (see UploadMultipartPart).

When you initiate a multipart upload, you specify the part size in number of bytes. The part size must be a megabyte (1024 KB) multiplied by a power of 2-for example, 1048576 (1 MB), 2097152 (2 MB), 4194304 (4 MB), 8388608 (8 MB), and so on. The minimum allowable part size is 1 MB, and the maximum is 4 GB.

Every part you upload to this resource (see UploadMultipartPart), except the last one, must have the same size. The last one can be the same size or smaller. For example, suppose you want to upload a 16.2 MB file. If you initiate the multipart upload with a part size of 4 MB, you will upload four parts of 4 MB each and one part of 0.2 MB.

You don't need to know the size of the archive when you start a multipart upload because Amazon Glacier does not require you to specify the overall archive size.

After you complete the multipart upload, Amazon Glacier removes the multipart upload resource referenced by the ID. Amazon Glacier also removes the multipart upload resource if you cancel the multipart upload or it may be removed if there is no activity for a period of 24 hours.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identityand Access Management (IAM).

For conceptual information and underlying REST API, go to Uploading LargeArchives in Parts (Multipart Upload) and Initiate Multipart Upload in the Amazon Glacier Developer Guide.

http://docs.aws.amazon.com/amazonglacier/latest/dev/api-InitiateMultipartUpload.html

Request

data InitiateMultipartUpload

Request constructor

initiateMultipartUpload

Arguments

:: Text
imuAccountId

-> Text
imuVaultName

-> InitiateMultipartUpload

InitiateMultipartUpload constructor.

The fields accessible through corresponding lenses are:

imuAccountId :: Text

imuArchiveDescription :: Maybe Text

imuPartSize :: Maybe Text

imuVaultName :: Text

Request lenses

imuAccountId :: Lens' InitiateMultipartUpload Text

The AccountId is the AWS Account ID. You can specify either the AWS Account ID or optionally a -, in which case Amazon Glacier uses the AWS Account ID associated with the credentials used to sign the request. If you specify your Account ID, do not include hyphens in it.

imuArchiveDescription :: Lens' InitiateMultipartUpload (Maybe Text)

The archive description that you are uploading in parts.

The part size must be a megabyte (1024 KB) multiplied by a power of 2—for example, 1048576 (1 MB), 2097152 (2 MB), 4194304 (4 MB), 8388608 (8 MB), and so on. The minimum allowable part size is 1 MB, and the maximum is 4 GB (4096 MB).

imuPartSize :: Lens' InitiateMultipartUpload (Maybe Text)

The size of each part except the last, in bytes. The last part can be smaller than this part size.

imuVaultName :: Lens' InitiateMultipartUpload Text

The name of the vault.

Response

data InitiateMultipartUploadResponse

Response constructor

initiateMultipartUploadResponse :: InitiateMultipartUploadResponse

InitiateMultipartUploadResponse constructor.

The fields accessible through corresponding lenses are:

imurLocation :: Maybe Text

imurUploadId :: Maybe Text

Response lenses

imurLocation :: Lens' InitiateMultipartUploadResponse (Maybe Text)

The relative URI path of the multipart upload ID Amazon Glacier created.

imurUploadId :: Lens' InitiateMultipartUploadResponse (Maybe Text)

The ID of the multipart upload. This value is also included as part of the location.