Network.AWS.Glacier.ListJobs

Description

This operation lists jobs for a vault, including jobs that are in-progress and jobs that have recently finished.

Amazon Glacier retains recently completed jobs for a period before deleting them; however, it eventually removes completed jobs. The output of completed jobs can be retrieved. Retaining completed jobs for a period of time after they have completed enables you to get a job output in the event you miss the job completion notification or your first attempt to download it fails. For example, suppose you start an archive retrieval job to download an archive. After the job completes, you start to download the archive but encounter a network error. In this scenario, you can retry and download the archive while the job exists.

To retrieve an archive or retrieve a vault inventory from Amazon Glacier, you first initiate a job, and after the job completes, you download the data. For an archive retrieval, the output is the archive data, and for an inventory retrieval, it is the inventory list. The List Job operation returns a list of these jobs sorted by job initiation time.

This List Jobs operation supports pagination. By default, this operation returns up to 1,000 jobs in the response. You should always check the response for a marker at which to continue the list; if there are no more items the marker is null. To return a list of jobs that begins at a specific job, set the marker request parameter to the value you obtained from a previous List Jobs request. You can also limit the number of jobs returned in the response by specifying the limit parameter in the request.

Additionally, you can filter the jobs list returned by specifying an optional statuscode (InProgress, Succeeded, or Failed) and completed (true, false) parameter. The statuscode allows you to specify that only jobs that match a specified status are returned. The completed parameter allows you to specify that only jobs in a specific completion state are returned.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identityand Access Management (IAM).

For the underlying REST API, go to List Jobs

http://docs.aws.amazon.com/amazonglacier/latest/dev/api-ListJobs.html

Request

data ListJobs

Request constructor

listJobs

Arguments

:: Text
ljAccountId

-> Text
ljVaultName

-> ListJobs

ListJobs constructor.

The fields accessible through corresponding lenses are:

ljAccountId :: Text

ljCompleted :: Maybe Text

ljLimit :: Maybe Text

ljMarker :: Maybe Text

ljStatuscode :: Maybe Text

ljVaultName :: Text

Request lenses

ljAccountId :: Lens' ListJobs Text

The AccountId is the AWS Account ID. You can specify either the AWS Account ID or optionally a -, in which case Amazon Glacier uses the AWS Account ID associated with the credentials used to sign the request. If you specify your Account ID, do not include hyphens in it.

ljCompleted :: Lens' ListJobs (Maybe Text)

Specifies the state of the jobs to return. You can specify true or false.

ljLimit :: Lens' ListJobs (Maybe Text)

Specifies that the response be limited to the specified number of items or fewer. If not specified, the List Jobs operation returns up to 1,000 jobs.

ljMarker :: Lens' ListJobs (Maybe Text)

An opaque string used for pagination. This value specifies the job at which the listing of jobs should begin. Get the marker value from a previous List Jobs response. You need only include the marker if you are continuing the pagination of results started in a previous List Jobs request.

ljStatuscode :: Lens' ListJobs (Maybe Text)

Specifies the type of job status to return. You can specify the following values: InProgress, Succeeded, or Failed.

ljVaultName :: Lens' ListJobs Text

The name of the vault.

Response

data ListJobsResponse

Response constructor

listJobsResponse :: ListJobsResponse

ListJobsResponse constructor.

The fields accessible through corresponding lenses are:

ljrJobList :: [GlacierJobDescription]

ljrMarker :: Maybe Text

Response lenses

ljrJobList :: Lens' ListJobsResponse [GlacierJobDescription]

A list of job objects. Each job object contains metadata describing the job.

ljrMarker :: Lens' ListJobsResponse (Maybe Text)

An opaque string that represents where to continue pagination of the results. You use this value in a new List Jobs request to obtain more jobs in the list. If there are no more jobs, this value is null.