Network.AWS.Glacier.ListMultipartUploads

Description

This operation lists in-progress multipart uploads for the specified vault. An in-progress multipart upload is a multipart upload that has been initiated by an InitiateMultipartUpload request, but has not yet been completed or aborted. The list returned in the List Multipart Upload response has no guaranteed order.

The List Multipart Uploads operation supports pagination. By default, this operation returns up to 1,000 multipart uploads in the response. You should always check the response for a marker at which to continue the list; if there are no more items the marker is null. To return a list of multipart uploads that begins at a specific upload, set the marker request parameter to the value you obtained from a previous List Multipart Upload request. You can also limit the number of uploads returned in the response by specifying the limit parameter in the request.

Note the difference between this operation and listing parts (ListParts). The List Multipart Uploads operation lists all multipart uploads for a vault and does not require a multipart upload ID. The List Parts operation requires a multipart upload ID since parts are associated with a single upload.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identityand Access Management (IAM).

For conceptual information and the underlying REST API, go to Working withArchives in Amazon Glacier and List Multipart Uploads in the Amazon GlacierDeveloper Guide.

http://docs.aws.amazon.com/amazonglacier/latest/dev/api-ListMultipartUploads.html

Request

data ListMultipartUploads

Request constructor

listMultipartUploads

Arguments

:: Text
lmuAccountId

-> Text
lmuVaultName

-> ListMultipartUploads

ListMultipartUploads constructor.

The fields accessible through corresponding lenses are:

lmuAccountId :: Text

lmuLimit :: Maybe Text

lmuMarker :: Maybe Text

lmuVaultName :: Text

Request lenses

lmuAccountId :: Lens' ListMultipartUploads Text

The AccountId is the AWS Account ID. You can specify either the AWS Account ID or optionally a -, in which case Amazon Glacier uses the AWS Account ID associated with the credentials used to sign the request. If you specify your Account ID, do not include hyphens in it.

lmuLimit :: Lens' ListMultipartUploads (Maybe Text)

Specifies the maximum number of uploads returned in the response body. If this value is not specified, the List Uploads operation returns up to 1,000 uploads.

lmuMarker :: Lens' ListMultipartUploads (Maybe Text)

An opaque string used for pagination. This value specifies the upload at which the listing of uploads should begin. Get the marker value from a previous List Uploads response. You need only include the marker if you are continuing the pagination of results started in a previous List Uploads request.

lmuVaultName :: Lens' ListMultipartUploads Text

The name of the vault.

Response

data ListMultipartUploadsResponse

Response constructor

listMultipartUploadsResponse :: ListMultipartUploadsResponse

ListMultipartUploadsResponse constructor.

The fields accessible through corresponding lenses are:

lmurMarker :: Maybe Text

lmurUploadsList :: [UploadListElement]

Response lenses

lmurMarker :: Lens' ListMultipartUploadsResponse (Maybe Text)

An opaque string that represents where to continue pagination of the results. You use the marker in a new List Multipart Uploads request to obtain more uploads in the list. If there are no more uploads, this value is null.

lmurUploadsList :: Lens' ListMultipartUploadsResponse [UploadListElement]

A list of in-progress multipart uploads.