Network.AWS.Glacier.ListParts

Description

This operation lists the parts of an archive that have been uploaded in a specific multipart upload. You can make this request at any time during an in-progress multipart upload before you complete the upload (see CompleteMultipartUpload. List Parts returns an error for completed uploads. The list returned in the List Parts response is sorted by part range.

The List Parts operation supports pagination. By default, this operation returns up to 1,000 uploaded parts in the response. You should always check the response for a marker at which to continue the list; if there are no more items the marker is null. To return a list of parts that begins at a specific part, set the marker request parameter to the value you obtained from a previous List Parts request. You can also limit the number of parts returned in the response by specifying the limit parameter in the request.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identityand Access Management (IAM).

For conceptual information and the underlying REST API, go to Working withArchives in Amazon Glacier and List Parts in the Amazon Glacier DeveloperGuide.

http://docs.aws.amazon.com/amazonglacier/latest/dev/api-ListParts.html

Request

data ListParts

Request constructor

listParts

Arguments

:: Text
lpAccountId

-> Text
lpVaultName

-> Text
lpUploadId

-> ListParts

ListParts constructor.

The fields accessible through corresponding lenses are:

lpAccountId :: Text

lpLimit :: Maybe Text

lpMarker :: Maybe Text

lpUploadId :: Text

lpVaultName :: Text

Request lenses

lpAccountId :: Lens' ListParts Text

The AccountId is the AWS Account ID. You can specify either the AWS Account ID or optionally a -, in which case Amazon Glacier uses the AWS Account ID associated with the credentials used to sign the request. If you specify your Account ID, do not include hyphens in it.

lpLimit :: Lens' ListParts (Maybe Text)

Specifies the maximum number of parts returned in the response body. If this value is not specified, the List Parts operation returns up to 1,000 uploads.

lpMarker :: Lens' ListParts (Maybe Text)

An opaque string used for pagination. This value specifies the part at which the listing of parts should begin. Get the marker value from the response of a previous List Parts response. You need only include the marker if you are continuing the pagination of results started in a previous List Parts request.

lpUploadId :: Lens' ListParts Text

The upload ID of the multipart upload.

lpVaultName :: Lens' ListParts Text

The name of the vault.

Response

data ListPartsResponse

Response constructor

listPartsResponse :: ListPartsResponse

ListPartsResponse constructor.

The fields accessible through corresponding lenses are:

lprArchiveDescription :: Maybe Text

lprCreationDate :: Maybe Text

lprMarker :: Maybe Text

lprMultipartUploadId :: Maybe Text

lprPartSizeInBytes :: Maybe Integer

lprParts :: [PartListElement]

lprVaultARN :: Maybe Text

Response lenses

lprArchiveDescription :: Lens' ListPartsResponse (Maybe Text)

The description of the archive that was specified in the Initiate Multipart Upload request.

lprCreationDate :: Lens' ListPartsResponse (Maybe Text)

The UTC time at which the multipart upload was initiated.

lprMarker :: Lens' ListPartsResponse (Maybe Text)

An opaque string that represents where to continue pagination of the results. You use the marker in a new List Parts request to obtain more jobs in the list. If there are no more parts, this value is null.

lprMultipartUploadId :: Lens' ListPartsResponse (Maybe Text)

The ID of the upload to which the parts are associated.

lprPartSizeInBytes :: Lens' ListPartsResponse (Maybe Integer)

The part size in bytes.

lprParts :: Lens' ListPartsResponse [PartListElement]

A list of the part sizes of the multipart upload.

lprVaultARN :: Lens' ListPartsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the vault to which the multipart upload was initiated.