Network.AWS.Glacier.ListVaults

Description

This operation lists all vaults owned by the calling user's account. The list returned in the response is ASCII-sorted by vault name.

By default, this operation returns up to 1,000 items. If there are more vaults to list, the response marker field contains the vault Amazon Resource Name (ARN) at which to continue the list with a new List Vaults request; otherwise, the marker field is null. To return a list of vaults that begins at a specific vault, set the marker request parameter to the vault ARN you obtained from a previous List Vaults request. You can also limit the number of vaults returned in the response by specifying the limit parameter in the request.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identityand Access Management (IAM).

For conceptual information and underlying REST API, go to Retrieving VaultMetadata in Amazon Glacier and List Vaults in the Amazon Glacier DeveloperGuide.

http://docs.aws.amazon.com/amazonglacier/latest/dev/api-ListVaults.html

Request

data ListVaults

Request constructor

listVaults

Arguments

:: Text
lvAccountId

-> ListVaults

ListVaults constructor.

The fields accessible through corresponding lenses are:

lvAccountId :: Text

lvLimit :: Maybe Text

lvMarker :: Maybe Text

Request lenses

lvAccountId :: Lens' ListVaults Text

The AccountId is the AWS Account ID. You can specify either the AWS Account ID or optionally a -, in which case Amazon Glacier uses the AWS Account ID associated with the credentials used to sign the request. If you specify your Account ID, do not include hyphens in it.

lvLimit :: Lens' ListVaults (Maybe Text)

The maximum number of items returned in the response. If you don't specify a value, the List Vaults operation returns up to 1,000 items.

lvMarker :: Lens' ListVaults (Maybe Text)

A string used for pagination. The marker specifies the vault ARN after which the listing of vaults should begin.

Response

data ListVaultsResponse

Response constructor

listVaultsResponse :: ListVaultsResponse

ListVaultsResponse constructor.

The fields accessible through corresponding lenses are:

lvrMarker :: Maybe Text

lvrVaultList :: [DescribeVaultOutput]

Response lenses

lvrMarker :: Lens' ListVaultsResponse (Maybe Text)

The vault ARN at which to continue pagination of the results. You use the marker in another List Vaults request to obtain more vaults in the list.

lvrVaultList :: Lens' ListVaultsResponse [DescribeVaultOutput]

List of vaults.