Network.AWS.Glacier.PurchaseProvisionedCapacity

Description

This operation purchases a provisioned capacity unit for an AWS account.

Creating a Request

purchaseProvisionedCapacity

Arguments

:: Text
ppcAccountId

-> PurchaseProvisionedCapacity

Creates a value of PurchaseProvisionedCapacity with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ppcAccountId - The AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single - (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, don't include any hyphens (-) in the ID.

data PurchaseProvisionedCapacity

See: purchaseProvisionedCapacity smart constructor.

Request Lenses

ppcAccountId :: Lens' PurchaseProvisionedCapacity Text

The AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single - (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, don't include any hyphens (-) in the ID.

Destructuring the Response

purchaseProvisionedCapacityResponse

Arguments

:: Int
ppcrsResponseStatus

-> PurchaseProvisionedCapacityResponse

Creates a value of PurchaseProvisionedCapacityResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ppcrsCapacityId - The ID that identifies the provisioned capacity unit.

ppcrsResponseStatus - -- | The response status code.

data PurchaseProvisionedCapacityResponse

See: purchaseProvisionedCapacityResponse smart constructor.

Response Lenses

ppcrsCapacityId :: Lens' PurchaseProvisionedCapacityResponse (Maybe Text)

The ID that identifies the provisioned capacity unit.

ppcrsResponseStatus :: Lens' PurchaseProvisionedCapacityResponse Int

- | The response status code.