Network.AWS.Glacier.RemoveTagsFromVault

Description

This operation removes one or more tags from the set of tags attached to a vault. For more information about tags, see Tagging Amazon Glacier Resources . This operation is idempotent. The operation will be successful, even if there are no tags attached to the vault.

Creating a Request

removeTagsFromVault

Arguments

:: Text
rtfvAccountId

-> Text
rtfvVaultName

-> RemoveTagsFromVault

Creates a value of RemoveTagsFromVault with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtfvTagKeys - A list of tag keys. Each corresponding tag is removed from the vault.

rtfvAccountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

rtfvVaultName - The name of the vault.

data RemoveTagsFromVault

The input value for RemoveTagsFromVaultInput .

See: removeTagsFromVault smart constructor.

Request Lenses

rtfvTagKeys :: Lens' RemoveTagsFromVault [Text]

A list of tag keys. Each corresponding tag is removed from the vault.

rtfvAccountId :: Lens' RemoveTagsFromVault Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

rtfvVaultName :: Lens' RemoveTagsFromVault Text

The name of the vault.

Destructuring the Response

removeTagsFromVaultResponse :: RemoveTagsFromVaultResponse

Creates a value of RemoveTagsFromVaultResponse with the minimum fields required to make a request.

data RemoveTagsFromVaultResponse

See: removeTagsFromVaultResponse smart constructor.