Network.AWS.Glacier.SetVaultAccessPolicy

Description

This operation configures an access policy for a vault and will overwrite an existing policy. To configure a vault access policy, send a PUT request to the access-policy subresource of the vault. An access policy is specific to a vault and is also called a vault subresource. You can set one access policy per vault and the policy can be up to 20 KB in size. For more information about vault access policies, see Amazon Glacier Access Control with Vault Access Policies .

Creating a Request

setVaultAccessPolicy

Arguments

:: Text
svapAccountId

-> Text
svapVaultName

-> SetVaultAccessPolicy

Creates a value of SetVaultAccessPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

svapPolicy - The vault access policy as a JSON string.

svapAccountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

svapVaultName - The name of the vault.

data SetVaultAccessPolicy

SetVaultAccessPolicy input.

See: setVaultAccessPolicy smart constructor.

Request Lenses

svapPolicy :: Lens' SetVaultAccessPolicy (Maybe VaultAccessPolicy)

The vault access policy as a JSON string.

svapAccountId :: Lens' SetVaultAccessPolicy Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

svapVaultName :: Lens' SetVaultAccessPolicy Text

The name of the vault.

Destructuring the Response

setVaultAccessPolicyResponse :: SetVaultAccessPolicyResponse

Creates a value of SetVaultAccessPolicyResponse with the minimum fields required to make a request.

data SetVaultAccessPolicyResponse

See: setVaultAccessPolicyResponse smart constructor.