Network.AWS.Glacier.SetVaultNotifications

Description

This operation configures notifications that will be sent when specific events happen to a vault. By default, you don't get any notifications.

To configure vault notifications, send a PUT request to the 'notification-configuration' subresource of the vault. The request should include a JSON document that provides an Amazon SNS topic and specific events for which you want Amazon Glacier to send notifications to the topic.

Amazon SNS topics must grant permission to the vault to be allowed to publish notifications to the topic. You can configure a vault to publish a notification for the following vault events:

ArchiveRetrievalCompleted This event occurs when a job that was initiated for an archive retrieval is completed (InitiateJob). The status of the completed job can be Succeeded or Failed. The notification sent to the SNS topic is the same output as returned from DescribeJob. InventoryRetrievalCompleted This event occurs when a job that was initiated for an inventory retrieval is completed (InitiateJob). The status of the completed job can be Succeeded or Failed. The notification sent to the SNS topic is the same output as returned from DescribeJob. An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identity and Access Management (IAM).

For conceptual information and underlying REST API, go to Configuring VaultNotifications in Amazon Glacier and Set Vault Notification Configuration in the Amazon Glacier Developer Guide.

http://docs.aws.amazon.com/amazonglacier/latest/dev/api-SetVaultNotifications.html

Request

data SetVaultNotifications

Request constructor

setVaultNotifications

Arguments

:: Text
svnAccountId

-> Text
svnVaultName

-> SetVaultNotifications

SetVaultNotifications constructor.

The fields accessible through corresponding lenses are:

svnAccountId :: Text

svnVaultName :: Text

svnVaultNotificationConfig :: Maybe VaultNotificationConfig

Request lenses

svnAccountId :: Lens' SetVaultNotifications Text

The AccountId is the AWS Account ID. You can specify either the AWS Account ID or optionally a -, in which case Amazon Glacier uses the AWS Account ID associated with the credentials used to sign the request. If you specify your Account ID, do not include hyphens in it.

svnVaultName :: Lens' SetVaultNotifications Text

The name of the vault.

svnVaultNotificationConfig :: Lens' SetVaultNotifications (Maybe VaultNotificationConfig)

Provides options for specifying notification configuration.

Response

data SetVaultNotificationsResponse

Response constructor

setVaultNotificationsResponse :: SetVaultNotificationsResponse

SetVaultNotificationsResponse constructor.