Network.AWS.Glacier.Types

Service

data Glacier

Version 2012-06-01 of the Amazon Glacier service.

Error

data JSONError :: *

ArchiveCreationOutput

data ArchiveCreationOutput

archiveCreationOutput :: ArchiveCreationOutput

ArchiveCreationOutput constructor.

The fields accessible through corresponding lenses are:

acoArchiveId :: Maybe Text

acoChecksum :: Maybe Text

acoLocation :: Maybe Text

acoArchiveId :: Lens' ArchiveCreationOutput (Maybe Text)

The ID of the archive. This value is also included as part of the location.

acoChecksum :: Lens' ArchiveCreationOutput (Maybe Text)

The checksum of the archive computed by Amazon Glacier.

acoLocation :: Lens' ArchiveCreationOutput (Maybe Text)

The relative URI path of the newly added archive resource.

UploadListElement

data UploadListElement

uploadListElement :: UploadListElement

UploadListElement constructor.

The fields accessible through corresponding lenses are:

uleArchiveDescription :: Maybe Text

uleCreationDate :: Maybe Text

uleMultipartUploadId :: Maybe Text

ulePartSizeInBytes :: Maybe Integer

uleVaultARN :: Maybe Text

uleArchiveDescription :: Lens' UploadListElement (Maybe Text)

The description of the archive that was specified in the Initiate Multipart Upload request.

uleCreationDate :: Lens' UploadListElement (Maybe Text)

The UTC time at which the multipart upload was initiated.

uleMultipartUploadId :: Lens' UploadListElement (Maybe Text)

The ID of a multipart upload.

ulePartSizeInBytes :: Lens' UploadListElement (Maybe Integer)

The part size, in bytes, specified in the Initiate Multipart Upload request. This is the size of all the parts in the upload except the last part, which may be smaller than this size.

uleVaultARN :: Lens' UploadListElement (Maybe Text)

The Amazon Resource Name (ARN) of the vault that contains the archive.

InventoryRetrievalJobDescription

data InventoryRetrievalJobDescription

inventoryRetrievalJobDescription :: InventoryRetrievalJobDescription

InventoryRetrievalJobDescription constructor.

The fields accessible through corresponding lenses are:

irjdEndDate :: Maybe Text

irjdFormat :: Maybe Text

irjdLimit :: Maybe Text

irjdMarker :: Maybe Text

irjdStartDate :: Maybe Text

irjdEndDate :: Lens' InventoryRetrievalJobDescription (Maybe Text)

The end of the date range in UTC for vault inventory retrieval that includes archives created before this date. A string representation of ISO 8601 date format, for example, 2013-03-20T17:03:43Z.

irjdFormat :: Lens' InventoryRetrievalJobDescription (Maybe Text)

The output format for the vault inventory list, which is set by the InitiateJob request when initiating a job to retrieve a vault inventory. Valid values are CSV and JSON.

irjdLimit :: Lens' InventoryRetrievalJobDescription (Maybe Text)

Specifies the maximum number of inventory items returned per vault inventory retrieval request. This limit is set when initiating the job with the a InitiateJob request.

irjdMarker :: Lens' InventoryRetrievalJobDescription (Maybe Text)

An opaque string that represents where to continue pagination of the vault inventory retrieval results. You use the marker in a new InitiateJob request to obtain additional inventory items. If there are no more inventory items, this value is null. For more information, see Range Inventory Retrieval.

irjdStartDate :: Lens' InventoryRetrievalJobDescription (Maybe Text)

The start of the date range in UTC for vault inventory retrieval that includes archives created on or after this date. A string representation of ISO 8601 date format, for example, 2013-03-20T17:03:43Z.

JobParameters

data JobParameters

jobParameters :: JobParameters

JobParameters constructor.

The fields accessible through corresponding lenses are:

jpArchiveId :: Maybe Text

jpDescription :: Maybe Text

jpFormat :: Maybe Text

jpInventoryRetrievalParameters :: Maybe InventoryRetrievalJobInput

jpRetrievalByteRange :: Maybe Text

jpSNSTopic :: Maybe Text

jpType :: Maybe Text

jpArchiveId :: Lens' JobParameters (Maybe Text)

The ID of the archive that you want to retrieve. This field is required only if Type is set to archive-retrieval. An error occurs if you specify this request parameter for an inventory retrieval job request.

jpDescription :: Lens' JobParameters (Maybe Text)

The optional description for the job. The description must be less than or equal to 1,024 bytes. The allowable characters are 7-bit ASCII without control codes-specifically, ASCII values 32-126 decimal or 0x20-0x7E hexadecimal.

jpFormat :: Lens' JobParameters (Maybe Text)

When initiating a job to retrieve a vault inventory, you can optionally add this parameter to your request to specify the output format. If you are initiating an inventory job and do not specify a Format field, JSON is the default format. Valid values are CSV and JSON.

jpInventoryRetrievalParameters :: Lens' JobParameters (Maybe InventoryRetrievalJobInput)

Input parameters used for range inventory retrieval.

jpRetrievalByteRange :: Lens' JobParameters (Maybe Text)

The byte range to retrieve for an archive retrieval. in the form "StartByteValue-EndByteValue" If not specified, the whole archive is retrieved. If specified, the byte range must be megabyte (1024*1024) aligned which means that StartByteValue must be divisible by 1 MB and EndByteValue plus 1 must be divisible by 1 MB or be the end of the archive specified as the archive byte size value minus 1. If RetrievalByteRange is not megabyte aligned, this operation returns a 400 response.

An error occurs if you specify this field for an inventory retrieval job request.

jpSNSTopic :: Lens' JobParameters (Maybe Text)

The Amazon SNS topic ARN to which Amazon Glacier sends a notification when the job is completed and the output is ready for you to download. The specified topic publishes the notification to its subscribers. The SNS topic must exist.

jpType :: Lens' JobParameters (Maybe Text)

The job type. You can initiate a job to retrieve an archive or get an inventory of a vault. Valid values are "archive-retrieval" and "inventory-retrieval".

DescribeVaultOutput

data DescribeVaultOutput

describeVaultOutput :: DescribeVaultOutput

DescribeVaultOutput constructor.

The fields accessible through corresponding lenses are:

dvoCreationDate :: Maybe Text

dvoLastInventoryDate :: Maybe Text

dvoNumberOfArchives :: Maybe Integer

dvoSizeInBytes :: Maybe Integer

dvoVaultARN :: Maybe Text

dvoVaultName :: Maybe Text

dvoCreationDate :: Lens' DescribeVaultOutput (Maybe Text)

The UTC date when the vault was created. A string representation of ISO 8601 date format, for example, "2012-03-20T17:03:43.221Z".

dvoLastInventoryDate :: Lens' DescribeVaultOutput (Maybe Text)

The UTC date when Amazon Glacier completed the last vault inventory. A string representation of ISO 8601 date format, for example, "2012-03-20T17:03:43.221Z".

dvoNumberOfArchives :: Lens' DescribeVaultOutput (Maybe Integer)

The number of archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example, if you just created the vault.

dvoSizeInBytes :: Lens' DescribeVaultOutput (Maybe Integer)

Total size, in bytes, of the archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example, if you just created the vault.

dvoVaultARN :: Lens' DescribeVaultOutput (Maybe Text)

The Amazon Resource Name (ARN) of the vault.

dvoVaultName :: Lens' DescribeVaultOutput (Maybe Text)

The name of the vault.

DataRetrievalRule

data DataRetrievalRule

dataRetrievalRule :: DataRetrievalRule

DataRetrievalRule constructor.

The fields accessible through corresponding lenses are:

drrBytesPerHour :: Maybe Integer

drrStrategy :: Maybe Text

drrBytesPerHour :: Lens' DataRetrievalRule (Maybe Integer)

drrStrategy :: Lens' DataRetrievalRule (Maybe Text)

ActionCode

data ActionCode

Constructors

ArchiveRetrieval
ArchiveRetrieval

InventoryRetrieval
InventoryRetrieval

VaultNotificationConfig

data VaultNotificationConfig

vaultNotificationConfig :: VaultNotificationConfig

VaultNotificationConfig constructor.

The fields accessible through corresponding lenses are:

vncEvents :: [Text]

vncSNSTopic :: Maybe Text

vncEvents :: Lens' VaultNotificationConfig [Text]

A list of one or more events for which Amazon Glacier will send a notification to the specified Amazon SNS topic.

vncSNSTopic :: Lens' VaultNotificationConfig (Maybe Text)

The Amazon Simple Notification Service (Amazon SNS) topic Amazon Resource Name (ARN).

InventoryRetrievalJobInput

data InventoryRetrievalJobInput

inventoryRetrievalJobInput :: InventoryRetrievalJobInput

InventoryRetrievalJobInput constructor.

The fields accessible through corresponding lenses are:

irjiEndDate :: Maybe Text

irjiLimit :: Maybe Text

irjiMarker :: Maybe Text

irjiStartDate :: Maybe Text

irjiEndDate :: Lens' InventoryRetrievalJobInput (Maybe Text)

The end of the date range in UTC for vault inventory retrieval that includes archives created before this date. A string representation of ISO 8601 date format, for example, 2013-03-20T17:03:43Z.

irjiLimit :: Lens' InventoryRetrievalJobInput (Maybe Text)

Specifies the maximum number of inventory items returned per vault inventory retrieval request. Valid values are greater than or equal to 1.

irjiMarker :: Lens' InventoryRetrievalJobInput (Maybe Text)

An opaque string that represents where to continue pagination of the vault inventory retrieval results. You use the marker in a new InitiateJob request to obtain additional inventory items. If there are no more inventory items, this value is null.

irjiStartDate :: Lens' InventoryRetrievalJobInput (Maybe Text)

The start of the date range in UTC for vault inventory retrieval that includes archives created on or after this date. A string representation of ISO 8601 date format, for example, 2013-03-20T17:03:43Z.

PartListElement

data PartListElement

partListElement :: PartListElement

PartListElement constructor.

The fields accessible through corresponding lenses are:

pleRangeInBytes :: Maybe Text

pleSHA256TreeHash :: Maybe Text

pleRangeInBytes :: Lens' PartListElement (Maybe Text)

The byte range of a part, inclusive of the upper value of the range.

pleSHA256TreeHash :: Lens' PartListElement (Maybe Text)

The SHA256 tree hash value that Amazon Glacier calculated for the part. This field is never null.

DataRetrievalPolicy

data DataRetrievalPolicy

dataRetrievalPolicy :: DataRetrievalPolicy

DataRetrievalPolicy constructor.

The fields accessible through corresponding lenses are:

drpRules :: [DataRetrievalRule]

drpRules :: Lens' DataRetrievalPolicy [DataRetrievalRule]

GlacierJobDescription

data GlacierJobDescription

glacierJobDescription :: GlacierJobDescription

GlacierJobDescription constructor.

The fields accessible through corresponding lenses are:

gjdAction :: Maybe ActionCode

gjdArchiveId :: Maybe Text

gjdArchiveSHA256TreeHash :: Maybe Text

gjdArchiveSizeInBytes :: Maybe Integer

gjdCompleted :: Maybe Bool

gjdCompletionDate :: Maybe Text

gjdCreationDate :: Maybe Text

gjdInventoryRetrievalParameters :: Maybe InventoryRetrievalJobDescription

gjdInventorySizeInBytes :: Maybe Integer

gjdJobDescription :: Maybe Text

gjdJobId :: Maybe Text

gjdRetrievalByteRange :: Maybe Text

gjdSHA256TreeHash :: Maybe Text

gjdSNSTopic :: Maybe Text

gjdStatusCode :: Maybe StatusCode

gjdStatusMessage :: Maybe Text

gjdVaultARN :: Maybe Text

gjdAction :: Lens' GlacierJobDescription (Maybe ActionCode)

The job type. It is either ArchiveRetrieval or InventoryRetrieval.

gjdArchiveId :: Lens' GlacierJobDescription (Maybe Text)

For an ArchiveRetrieval job, this is the archive ID requested for download. Otherwise, this field is null.

gjdArchiveSHA256TreeHash :: Lens' GlacierJobDescription (Maybe Text)

The SHA256 tree hash of the entire archive for an archive retrieval. For inventory retrieval jobs, this field is null.

gjdArchiveSizeInBytes :: Lens' GlacierJobDescription (Maybe Integer)

For an ArchiveRetrieval job, this is the size in bytes of the archive being requested for download. For the InventoryRetrieval job, the value is null.

gjdCompleted :: Lens' GlacierJobDescription (Maybe Bool)

The job status. When a job is completed, you get the job's output.

gjdCompletionDate :: Lens' GlacierJobDescription (Maybe Text)

The UTC time that the archive retrieval request completed. While the job is in progress, the value will be null.

gjdCreationDate :: Lens' GlacierJobDescription (Maybe Text)

The UTC date when the job was created. A string representation of ISO 8601 date format, for example, "2012-03-20T17:03:43.221Z".

gjdInventoryRetrievalParameters :: Lens' GlacierJobDescription (Maybe InventoryRetrievalJobDescription)

Parameters used for range inventory retrieval.

gjdInventorySizeInBytes :: Lens' GlacierJobDescription (Maybe Integer)

For an InventoryRetrieval job, this is the size in bytes of the inventory requested for download. For the ArchiveRetrieval job, the value is null.

gjdJobDescription :: Lens' GlacierJobDescription (Maybe Text)

The job description you provided when you initiated the job.

gjdJobId :: Lens' GlacierJobDescription (Maybe Text)

An opaque string that identifies an Amazon Glacier job.

gjdRetrievalByteRange :: Lens' GlacierJobDescription (Maybe Text)

The retrieved byte range for archive retrieval jobs in the form "StartByteValue-EndByteValue" If no range was specified in the archive retrieval, then the whole archive is retrieved and StartByteValue equals 0 and EndByteValue equals the size of the archive minus 1. For inventory retrieval jobs this field is null.

gjdSHA256TreeHash :: Lens' GlacierJobDescription (Maybe Text)

For an ArchiveRetrieval job, it is the checksum of the archive. Otherwise, the value is null.

The SHA256 tree hash value for the requested range of an archive. If the Initiate a Job request for an archive specified a tree-hash aligned range, then this field returns a value.

For the specific case when the whole archive is retrieved, this value is the same as the ArchiveSHA256TreeHash value.

This field is null in the following situations: Archive retrieval jobs that specify a range that is not tree-hash aligned.

Archival jobs that specify a range that is equal to the whole archive and the job status is InProgress.

Inventory jobs.

gjdSNSTopic :: Lens' GlacierJobDescription (Maybe Text)

An Amazon Simple Notification Service (Amazon SNS) topic that receives notification.

gjdStatusCode :: Lens' GlacierJobDescription (Maybe StatusCode)

The status code can be InProgress, Succeeded, or Failed, and indicates the status of the job.

gjdStatusMessage :: Lens' GlacierJobDescription (Maybe Text)

A friendly message that describes the job status.

gjdVaultARN :: Lens' GlacierJobDescription (Maybe Text)

The Amazon Resource Name (ARN) of the vault from which the archive retrieval was requested.

StatusCode

data StatusCode

Constructors

Failed
Failed

InProgress
InProgress

Succeeded
Succeeded