Network.AWS.Glacier.UploadArchive

Description

This operation adds an archive to a vault. This is a synchronous operation, and for a successful upload, your data is durably persisted. Amazon Glacier returns the archive ID in the 'x-amz-archive-id' header of the response.

You must use the archive ID to access your data in Amazon Glacier. After you upload an archive, you should save the archive ID returned so that you can retrieve or delete the archive later. Besides saving the archive ID, you can also index it and give it a friendly name to allow for better searching. You can also use the optional archive description field to specify how the archive is referred to in an external index of archives, such as you might create in Amazon DynamoDB. You can also get the vault inventory to obtain a list of archive IDs in a vault. For more information, see InitiateJob.

You must provide a SHA256 tree hash of the data you are uploading. For information about computing a SHA256 tree hash, see Computing Checksums.

You can optionally specify an archive description of up to 1,024 printable ASCII characters. You can get the archive description when you either retrieve the archive or get the vault inventory. For more information, see InitiateJob. Amazon Glacier does not interpret the description in any way. An archive description does not need to be unique. You cannot use the description to retrieve or sort the archive list.

Archives are immutable. After you upload an archive, you cannot edit the archive or its description.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identityand Access Management (IAM).

For conceptual information and underlying REST API, go to Uploading anArchive in Amazon Glacier and Upload Archive in the Amazon Glacier DeveloperGuide.

http://docs.aws.amazon.com/amazonglacier/latest/dev/api-UploadArchive.html

Request

data UploadArchive

Request constructor

uploadArchive

Arguments

:: Text
uaVaultName

-> Text
uaAccountId

-> RqBody
uaBody

-> UploadArchive

UploadArchive constructor.

The fields accessible through corresponding lenses are:

uaAccountId :: Text

uaArchiveDescription :: Maybe Text

uaBody :: RqBody

uaChecksum :: Maybe Text

uaVaultName :: Text

Request lenses

uaAccountId :: Lens' UploadArchive Text

The AccountId is the AWS Account ID. You can specify either the AWS Account ID or optionally a -, in which case Amazon Glacier uses the AWS Account ID associated with the credentials used to sign the request. If you specify your Account ID, do not include hyphens in it.

uaArchiveDescription :: Lens' UploadArchive (Maybe Text)

The optional description of the archive you are uploading.

uaBody :: Lens' UploadArchive RqBody

The data to upload.

uaChecksum :: Lens' UploadArchive (Maybe Text)

The SHA256 checksum (a linear hash) of the payload.

uaVaultName :: Lens' UploadArchive Text

The name of the vault.

Response

data UploadArchiveResponse

Response constructor

uploadArchiveResponse :: UploadArchiveResponse

UploadArchiveResponse constructor.

The fields accessible through corresponding lenses are:

uarArchiveId :: Maybe Text

uarChecksum :: Maybe Text

uarLocation :: Maybe Text

Response lenses

uarArchiveId :: Lens' UploadArchiveResponse (Maybe Text)

The ID of the archive. This value is also included as part of the location.

uarChecksum :: Lens' UploadArchiveResponse (Maybe Text)

The checksum of the archive computed by Amazon Glacier.

uarLocation :: Lens' UploadArchiveResponse (Maybe Text)

The relative URI path of the newly added archive resource.