Network.AWS.Glacier.UploadMultipartPart

Description

This operation uploads a part of an archive. You can upload archive parts in any order. You can also upload them in parallel. You can upload up to 10,000 parts for a multipart upload.

Amazon Glacier rejects your upload part request if any of the following conditions is true:

SHA256 tree hash does not matchTo ensure that part data is not corrupted in transmission, you compute a SHA256 tree hash of the part and include it in your request. Upon receiving the part data, Amazon Glacier also computes a SHA256 tree hash. If these hash values don't match, the operation fails. For information about computing a SHA256 tree hash, see Computing Checksums.

Part size does not matchThe size of each part except the last must match the size specified in the corresponding InitiateMultipartUpload request. The size of the last part must be the same size as, or smaller than, the specified size.

If you upload a part whose size is smaller than the part size you specified in your initiate multipart upload request and that part is not the last part, then the upload part request will succeed. However, the subsequent Complete Multipart Upload request will fail.

Range does not alignThe byte range value in the request does not align with the part size specified in the corresponding initiate request. For example, if you specify a part size of 4194304 bytes (4 MB), then 0 to 4194303 bytes (4 MB - 1) and 4194304 (4 MB) to 8388607 (8 MB - 1) are valid part ranges. However, if you set a range value of 2 MB to 6 MB, the range does not align with the part size and the upload will fail. This operation is idempotent. If you upload the same part multiple times, the data included in the most recent request overwrites the previously uploaded data.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identityand Access Management (IAM).

For conceptual information and underlying REST API, go to Uploading LargeArchives in Parts (Multipart Upload) and Upload Part in the Amazon GlacierDeveloper Guide.

http://docs.aws.amazon.com/amazonglacier/latest/dev/api-UploadMultipartPart.html

Request

data UploadMultipartPart

Request constructor

uploadMultipartPart

Arguments

:: Text
umpAccountId

-> Text
umpVaultName

-> Text
umpUploadId

-> RqBody
umpBody

-> UploadMultipartPart

UploadMultipartPart constructor.

The fields accessible through corresponding lenses are:

umpAccountId :: Text

umpBody :: RqBody

umpChecksum :: Maybe Text

umpRange :: Maybe Text

umpUploadId :: Text

umpVaultName :: Text

Request lenses

umpAccountId :: Lens' UploadMultipartPart Text

The AccountId is the AWS Account ID. You can specify either the AWS Account ID or optionally a -, in which case Amazon Glacier uses the AWS Account ID associated with the credentials used to sign the request. If you specify your Account ID, do not include hyphens in it.

umpBody :: Lens' UploadMultipartPart RqBody

The data to upload.

umpChecksum :: Lens' UploadMultipartPart (Maybe Text)

The SHA256 tree hash of the data being uploaded.

umpRange :: Lens' UploadMultipartPart (Maybe Text)

Identifies the range of bytes in the assembled archive that will be uploaded in this part. Amazon Glacier uses this information to assemble the archive in the proper sequence. The format of this header follows RFC 2616. An example header is Content-Range:bytes 0-4194303/*.

umpUploadId :: Lens' UploadMultipartPart Text

The upload ID of the multipart upload.

umpVaultName :: Lens' UploadMultipartPart Text

The name of the vault.

Response

data UploadMultipartPartResponse

Response constructor

uploadMultipartPartResponse :: UploadMultipartPartResponse

UploadMultipartPartResponse constructor.

The fields accessible through corresponding lenses are:

umprChecksum :: Maybe Text

Response lenses

umprChecksum :: Lens' UploadMultipartPartResponse (Maybe Text)

The SHA256 tree hash that Amazon Glacier computed for the uploaded part.