Network.AWS.Glacier.AbortMultipartUpload

Description

This operation aborts a multipart upload identified by the upload ID.

After the Abort Multipart Upload request succeeds, you cannot upload any more parts to the multipart upload or complete the multipart upload. Aborting a completed upload fails. However, aborting an already-aborted upload will succeed, for a short time. For more information about uploading a part and completing a multipart upload, see UploadMultipartPart and CompleteMultipartUpload .

This operation is idempotent.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identity and Access Management (IAM) .

For conceptual information and underlying REST API, see Working with Archives in Amazon Glacier and Abort Multipart Upload in the Amazon Glacier Developer Guide .

Creating a Request

abortMultipartUpload

Arguments

:: Text
amuAccountId

-> Text
amuVaultName

-> Text
amuUploadId

-> AbortMultipartUpload

Creates a value of AbortMultipartUpload with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

amuAccountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

amuVaultName - The name of the vault.

amuUploadId - The upload ID of the multipart upload to delete.

data AbortMultipartUpload

Provides options to abort a multipart upload identified by the upload ID.

For information about the underlying REST API, see Abort Multipart Upload . For conceptual information, see Working with Archives in Amazon Glacier .

See: abortMultipartUpload smart constructor.

Request Lenses

amuAccountId :: Lens' AbortMultipartUpload Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

amuVaultName :: Lens' AbortMultipartUpload Text

The name of the vault.

amuUploadId :: Lens' AbortMultipartUpload Text

The upload ID of the multipart upload to delete.

Destructuring the Response

abortMultipartUploadResponse :: AbortMultipartUploadResponse

Creates a value of AbortMultipartUploadResponse with the minimum fields required to make a request.

data AbortMultipartUploadResponse

See: abortMultipartUploadResponse smart constructor.