Network.AWS.Glacier.CreateVault

Description

This operation creates a new vault with the specified name. The name of the vault must be unique within a region for an AWS account. You can create up to 1,000 vaults per account. If you need to create more vaults, contact Amazon Glacier.

You must use the following guidelines when naming a vault.

Names can be between 1 and 255 characters long.

Allowed characters are a-z, A-Z, 0-9, '_' (underscore), - (hyphen), and . (period).

This operation is idempotent.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identity and Access Management (IAM) .

For conceptual information and underlying REST API, see Creating a Vault in Amazon Glacier and Create Vault in the Amazon Glacier Developer Guide .

Creating a Request

createVault

Arguments

:: Text
cvAccountId

-> Text
cvVaultName

-> CreateVault

Creates a value of CreateVault with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cvAccountId - The AccountId value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens (-) in the ID.

cvVaultName - The name of the vault.

data CreateVault

Provides options to create a vault.

See: createVault smart constructor.

Request Lenses

cvAccountId :: Lens' CreateVault Text

The AccountId value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens (-) in the ID.

cvVaultName :: Lens' CreateVault Text

The name of the vault.

Destructuring the Response

createVaultResponse

Arguments

:: Int
cvrsResponseStatus

-> CreateVaultResponse

Creates a value of CreateVaultResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cvrsLocation - The URI of the vault that was created.

cvrsResponseStatus - -- | The response status code.

data CreateVaultResponse

Contains the Amazon Glacier response to your request.

See: createVaultResponse smart constructor.

Response Lenses

cvrsLocation :: Lens' CreateVaultResponse (Maybe Text)

The URI of the vault that was created.

cvrsResponseStatus :: Lens' CreateVaultResponse Int

- | The response status code.