Network.AWS.Glacier.DescribeJob

Description

This operation returns information about a job you previously initiated, including the job initiation date, the user who initiated the job, the job status code/message and the Amazon SNS topic to notify after Amazon Glacier completes the job. For more information about initiating a job, see InitiateJob .

A job ID will not expire for at least 24 hours after Amazon Glacier completes the job.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identity and Access Management (IAM) .

For information about the underlying REST API, see Working with Archives in Amazon Glacier in the Amazon Glacier Developer Guide .

Creating a Request

describeJob

Arguments

:: Text
djAccountId

-> Text
djVaultName

-> Text
djJobId

-> DescribeJob

Creates a value of DescribeJob with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

djAccountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

djVaultName - The name of the vault.

djJobId - The ID of the job to describe.

data DescribeJob

Provides options for retrieving a job description.

See: describeJob smart constructor.

Request Lenses

djAccountId :: Lens' DescribeJob Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

djVaultName :: Lens' DescribeJob Text

The name of the vault.

djJobId :: Lens' DescribeJob Text

The ID of the job to describe.

Destructuring the Response

glacierJobDescription :: GlacierJobDescription

Creates a value of GlacierJobDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gjdSHA256TreeHash - For an ArchiveRetrieval job, it is the checksum of the archive. Otherwise, the value is null. The SHA256 tree hash value for the requested range of an archive. If the Initiate a Job request for an archive specified a tree-hash aligned range, then this field returns a value. For the specific case when the whole archive is retrieved, this value is the same as the ArchiveSHA256TreeHash value. This field is null in the following situations: * Archive retrieval jobs that specify a range that is not tree-hash aligned. * Archival jobs that specify a range that is equal to the whole archive and the job status is InProgress. * Inventory jobs.

gjdArchiveId - For an ArchiveRetrieval job, this is the archive ID requested for download. Otherwise, this field is null.

gjdJobId - An opaque string that identifies an Amazon Glacier job.

gjdRetrievalByteRange - The retrieved byte range for archive retrieval jobs in the form "StartByteValue -EndByteValue " If no range was specified in the archive retrieval, then the whole archive is retrieved and StartByteValue equals 0 and EndByteValue equals the size of the archive minus 1. For inventory retrieval jobs this field is null.

gjdInventoryRetrievalParameters - Parameters used for range inventory retrieval.

gjdAction - The job type. It is either ArchiveRetrieval or InventoryRetrieval.

gjdJobDescription - The job description you provided when you initiated the job.

gjdSNSTopic - An Amazon Simple Notification Service (Amazon SNS) topic that receives notification.

gjdStatusMessage - A friendly message that describes the job status.

gjdVaultARN - The Amazon Resource Name (ARN) of the vault from which the archive retrieval was requested.

gjdTier - The retrieval option to use for the archive retrieval. Valid values are Expedited , Standard , or Bulk . Standard is the default.

gjdArchiveSHA256TreeHash - The SHA256 tree hash of the entire archive for an archive retrieval. For inventory retrieval jobs, this field is null.

gjdCreationDate - The UTC date when the job was created. A string representation of ISO 8601 date format, for example, "2012-03-20T17:03:43.221Z".

gjdCompleted - The job status. When a job is completed, you get the job's output.

gjdCompletionDate - The UTC time that the archive retrieval request completed. While the job is in progress, the value will be null.

gjdInventorySizeInBytes - For an InventoryRetrieval job, this is the size in bytes of the inventory requested for download. For the ArchiveRetrieval job, the value is null.

gjdArchiveSizeInBytes - For an ArchiveRetrieval job, this is the size in bytes of the archive being requested for download. For the InventoryRetrieval job, the value is null.

gjdStatusCode - The status code can be InProgress, Succeeded, or Failed, and indicates the status of the job.

data GlacierJobDescription

Describes an Amazon Glacier job.

See: glacierJobDescription smart constructor.

Response Lenses

gjdSHA256TreeHash :: Lens' GlacierJobDescription (Maybe Text)

For an ArchiveRetrieval job, it is the checksum of the archive. Otherwise, the value is null. The SHA256 tree hash value for the requested range of an archive. If the Initiate a Job request for an archive specified a tree-hash aligned range, then this field returns a value. For the specific case when the whole archive is retrieved, this value is the same as the ArchiveSHA256TreeHash value. This field is null in the following situations: * Archive retrieval jobs that specify a range that is not tree-hash aligned. * Archival jobs that specify a range that is equal to the whole archive and the job status is InProgress. * Inventory jobs.

gjdArchiveId :: Lens' GlacierJobDescription (Maybe Text)

For an ArchiveRetrieval job, this is the archive ID requested for download. Otherwise, this field is null.

gjdJobId :: Lens' GlacierJobDescription (Maybe Text)

An opaque string that identifies an Amazon Glacier job.

gjdRetrievalByteRange :: Lens' GlacierJobDescription (Maybe Text)

The retrieved byte range for archive retrieval jobs in the form "StartByteValue -EndByteValue " If no range was specified in the archive retrieval, then the whole archive is retrieved and StartByteValue equals 0 and EndByteValue equals the size of the archive minus 1. For inventory retrieval jobs this field is null.

gjdInventoryRetrievalParameters :: Lens' GlacierJobDescription (Maybe InventoryRetrievalJobDescription)

Parameters used for range inventory retrieval.

gjdAction :: Lens' GlacierJobDescription (Maybe ActionCode)

The job type. It is either ArchiveRetrieval or InventoryRetrieval.

gjdJobDescription :: Lens' GlacierJobDescription (Maybe Text)

The job description you provided when you initiated the job.

gjdSNSTopic :: Lens' GlacierJobDescription (Maybe Text)

An Amazon Simple Notification Service (Amazon SNS) topic that receives notification.

gjdStatusMessage :: Lens' GlacierJobDescription (Maybe Text)

A friendly message that describes the job status.

gjdVaultARN :: Lens' GlacierJobDescription (Maybe Text)

The Amazon Resource Name (ARN) of the vault from which the archive retrieval was requested.

gjdTier :: Lens' GlacierJobDescription (Maybe Text)

The retrieval option to use for the archive retrieval. Valid values are Expedited , Standard , or Bulk . Standard is the default.

gjdArchiveSHA256TreeHash :: Lens' GlacierJobDescription (Maybe Text)

The SHA256 tree hash of the entire archive for an archive retrieval. For inventory retrieval jobs, this field is null.

gjdCreationDate :: Lens' GlacierJobDescription (Maybe Text)

The UTC date when the job was created. A string representation of ISO 8601 date format, for example, "2012-03-20T17:03:43.221Z".

gjdCompleted :: Lens' GlacierJobDescription (Maybe Bool)

The job status. When a job is completed, you get the job's output.

gjdCompletionDate :: Lens' GlacierJobDescription (Maybe Text)

The UTC time that the archive retrieval request completed. While the job is in progress, the value will be null.

gjdInventorySizeInBytes :: Lens' GlacierJobDescription (Maybe Integer)

For an InventoryRetrieval job, this is the size in bytes of the inventory requested for download. For the ArchiveRetrieval job, the value is null.

gjdArchiveSizeInBytes :: Lens' GlacierJobDescription (Maybe Integer)

For an ArchiveRetrieval job, this is the size in bytes of the archive being requested for download. For the InventoryRetrieval job, the value is null.

gjdStatusCode :: Lens' GlacierJobDescription (Maybe StatusCode)

The status code can be InProgress, Succeeded, or Failed, and indicates the status of the job.