Network.AWS.Glacier.DescribeVault

Description

This operation returns information about a vault, including the vault's Amazon Resource Name (ARN), the date the vault was created, the number of archives it contains, and the total size of all the archives in the vault. The number of archives and their total size are as of the last inventory generation. This means that if you add or remove an archive from a vault, and then immediately use Describe Vault, the change in contents will not be immediately reflected. If you want to retrieve the latest inventory of the vault, use InitiateJob . Amazon Glacier generates vault inventories approximately daily. For more information, see Downloading a Vault Inventory in Amazon Glacier .

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identity and Access Management (IAM) .

For conceptual information and underlying REST API, see Retrieving Vault Metadata in Amazon Glacier and Describe Vault in the Amazon Glacier Developer Guide .

Creating a Request

describeVault

Arguments

:: Text
dvAccountId

-> Text
dvVaultName

-> DescribeVault

Creates a value of DescribeVault with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvAccountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

dvVaultName - The name of the vault.

data DescribeVault

Provides options for retrieving metadata for a specific vault in Amazon Glacier.

See: describeVault smart constructor.

Request Lenses

dvAccountId :: Lens' DescribeVault Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

dvVaultName :: Lens' DescribeVault Text

The name of the vault.

Destructuring the Response

describeVaultOutput :: DescribeVaultOutput

Creates a value of DescribeVaultOutput with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvoVaultName - The name of the vault.

dvoSizeInBytes - Total size, in bytes, of the archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.

dvoLastInventoryDate - The Universal Coordinated Time (UTC) date when Amazon Glacier completed the last vault inventory. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z .

dvoVaultARN - The Amazon Resource Name (ARN) of the vault.

dvoCreationDate - The Universal Coordinated Time (UTC) date when the vault was created. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z .

dvoNumberOfArchives - The number of archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.

data DescribeVaultOutput

Contains the Amazon Glacier response to your request.

See: describeVaultOutput smart constructor.

Response Lenses

dvoVaultName :: Lens' DescribeVaultOutput (Maybe Text)

The name of the vault.

dvoSizeInBytes :: Lens' DescribeVaultOutput (Maybe Integer)

Total size, in bytes, of the archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.

dvoLastInventoryDate :: Lens' DescribeVaultOutput (Maybe Text)

The Universal Coordinated Time (UTC) date when Amazon Glacier completed the last vault inventory. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z .

dvoVaultARN :: Lens' DescribeVaultOutput (Maybe Text)

The Amazon Resource Name (ARN) of the vault.

dvoCreationDate :: Lens' DescribeVaultOutput (Maybe Text)

The Universal Coordinated Time (UTC) date when the vault was created. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z .

dvoNumberOfArchives :: Lens' DescribeVaultOutput (Maybe Integer)

The number of archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.