Network.AWS.Glacier.InitiateMultipartUpload

Description

This operation initiates a multipart upload. Amazon Glacier creates a multipart upload resource and returns its ID in the response. The multipart upload ID is used in subsequent requests to upload parts of an archive (see UploadMultipartPart ).

When you initiate a multipart upload, you specify the part size in number of bytes. The part size must be a megabyte (1024 KB) multiplied by a power of 2-for example, 1048576 (1 MB), 2097152 (2 MB), 4194304 (4 MB), 8388608 (8 MB), and so on. The minimum allowable part size is 1 MB, and the maximum is 4 GB.

Every part you upload to this resource (see UploadMultipartPart ), except the last one, must have the same size. The last one can be the same size or smaller. For example, suppose you want to upload a 16.2 MB file. If you initiate the multipart upload with a part size of 4 MB, you will upload four parts of 4 MB each and one part of 0.2 MB.

After you complete the multipart upload, Amazon Glacier removes the multipart upload resource referenced by the ID. Amazon Glacier also removes the multipart upload resource if you cancel the multipart upload or it may be removed if there is no activity for a period of 24 hours.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identity and Access Management (IAM) .

For conceptual information and underlying REST API, see Uploading Large Archives in Parts (Multipart Upload) and Initiate Multipart Upload in the Amazon Glacier Developer Guide .

Creating a Request

initiateMultipartUpload

Arguments

:: Text
imuAccountId

-> Text
imuVaultName

-> InitiateMultipartUpload

Creates a value of InitiateMultipartUpload with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

imuPartSize - The size of each part except the last, in bytes. The last part can be smaller than this part size.

imuArchiveDescription - The archive description that you are uploading in parts. The part size must be a megabyte (1024 KB) multiplied by a power of 2, for example 1048576 (1 MB), 2097152 (2 MB), 4194304 (4 MB), 8388608 (8 MB), and so on. The minimum allowable part size is 1 MB, and the maximum is 4 GB (4096 MB).

imuAccountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

imuVaultName - The name of the vault.

data InitiateMultipartUpload

Provides options for initiating a multipart upload to an Amazon Glacier vault.

See: initiateMultipartUpload smart constructor.

Request Lenses

imuPartSize :: Lens' InitiateMultipartUpload (Maybe Text)

The size of each part except the last, in bytes. The last part can be smaller than this part size.

imuArchiveDescription :: Lens' InitiateMultipartUpload (Maybe Text)

The archive description that you are uploading in parts. The part size must be a megabyte (1024 KB) multiplied by a power of 2, for example 1048576 (1 MB), 2097152 (2 MB), 4194304 (4 MB), 8388608 (8 MB), and so on. The minimum allowable part size is 1 MB, and the maximum is 4 GB (4096 MB).

imuAccountId :: Lens' InitiateMultipartUpload Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

imuVaultName :: Lens' InitiateMultipartUpload Text

The name of the vault.

Destructuring the Response

initiateMultipartUploadResponse

Arguments

:: Int
imursResponseStatus

-> InitiateMultipartUploadResponse

Creates a value of InitiateMultipartUploadResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

imursLocation - The relative URI path of the multipart upload ID Amazon Glacier created.

imursUploadId - The ID of the multipart upload. This value is also included as part of the location.

imursResponseStatus - -- | The response status code.

data InitiateMultipartUploadResponse

The Amazon Glacier response to your request.

See: initiateMultipartUploadResponse smart constructor.

Response Lenses

imursLocation :: Lens' InitiateMultipartUploadResponse (Maybe Text)

The relative URI path of the multipart upload ID Amazon Glacier created.

imursUploadId :: Lens' InitiateMultipartUploadResponse (Maybe Text)

The ID of the multipart upload. This value is also included as part of the location.

imursResponseStatus :: Lens' InitiateMultipartUploadResponse Int

- | The response status code.