Network.AWS.Glacier.ListJobs

Description

This operation lists jobs for a vault, including jobs that are in-progress and jobs that have recently finished.

To retrieve an archive or retrieve a vault inventory from Amazon Glacier, you first initiate a job, and after the job completes, you download the data. For an archive retrieval, the output is the archive data. For an inventory retrieval, it is the inventory list. The List Job operation returns a list of these jobs sorted by job initiation time.

The List Jobs operation supports pagination. You should always check the response Marker field. If there are no more jobs to list, the Marker field is set to null . If there are more jobs to list, the Marker field is set to a non-null value, which you can use to continue the pagination of the list. To return a list of jobs that begins at a specific job, set the marker request parameter to the Marker value for that job that you obtained from a previous List Jobs request.

You can set a maximum limit for the number of jobs returned in the response by specifying the limit parameter in the request. The default limit is 1000. The number of jobs returned might be fewer than the limit, but the number of returned jobs never exceeds the limit.

Additionally, you can filter the jobs list returned by specifying the optional statuscode parameter or completed parameter, or both. Using the statuscode parameter, you can specify to return only jobs that match either the InProgress , Succeeded , or Failed status. Using the completed parameter, you can specify to return only jobs that were completed (true ) or jobs that were not completed (false ).

For the underlying REST API, see List Jobs .

This operation returns paginated results.

Creating a Request

listJobs

Arguments

:: Text
ljAccountId

-> Text
ljVaultName

-> ListJobs

Creates a value of ListJobs with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ljMarker - An opaque string used for pagination. This value specifies the job at which the listing of jobs should begin. Get the marker value from a previous List Jobs response. You only need to include the marker if you are continuing the pagination of results started in a previous List Jobs request.

ljCompleted - The state of the jobs to return. You can specify true or false .

ljLimit - The maximum number of jobs to be returned. The default limit is 1000. The number of jobs returned might be fewer than the specified limit, but the number of returned jobs never exceeds the limit.

ljStatuscode - The type of job status to return. You can specify the following values: InProgress , Succeeded , or Failed .

ljAccountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

ljVaultName - The name of the vault.

data ListJobs

Provides options for retrieving a job list for an Amazon Glacier vault.

See: listJobs smart constructor.

Request Lenses

ljMarker :: Lens' ListJobs (Maybe Text)

An opaque string used for pagination. This value specifies the job at which the listing of jobs should begin. Get the marker value from a previous List Jobs response. You only need to include the marker if you are continuing the pagination of results started in a previous List Jobs request.

ljCompleted :: Lens' ListJobs (Maybe Text)

The state of the jobs to return. You can specify true or false .

ljLimit :: Lens' ListJobs (Maybe Text)

The maximum number of jobs to be returned. The default limit is 1000. The number of jobs returned might be fewer than the specified limit, but the number of returned jobs never exceeds the limit.

ljStatuscode :: Lens' ListJobs (Maybe Text)

The type of job status to return. You can specify the following values: InProgress , Succeeded , or Failed .

ljAccountId :: Lens' ListJobs Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

ljVaultName :: Lens' ListJobs Text

The name of the vault.

Destructuring the Response

listJobsResponse

Arguments

:: Int
ljrsResponseStatus

-> ListJobsResponse

Creates a value of ListJobsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ljrsMarker - An opaque string used for pagination that specifies the job at which the listing of jobs should begin. You get the marker value from a previous List Jobs response. You only need to include the marker if you are continuing the pagination of the results started in a previous List Jobs request.

ljrsJobList - A list of job objects. Each job object contains metadata describing the job.

ljrsResponseStatus - -- | The response status code.

data ListJobsResponse

Contains the Amazon Glacier response to your request.

See: listJobsResponse smart constructor.

Response Lenses

ljrsMarker :: Lens' ListJobsResponse (Maybe Text)

An opaque string used for pagination that specifies the job at which the listing of jobs should begin. You get the marker value from a previous List Jobs response. You only need to include the marker if you are continuing the pagination of the results started in a previous List Jobs request.

ljrsJobList :: Lens' ListJobsResponse [GlacierJobDescription]

A list of job objects. Each job object contains metadata describing the job.

ljrsResponseStatus :: Lens' ListJobsResponse Int

- | The response status code.