Network.AWS.Glacier.ListParts

Description

This operation lists the parts of an archive that have been uploaded in a specific multipart upload. You can make this request at any time during an in-progress multipart upload before you complete the upload (see CompleteMultipartUpload . List Parts returns an error for completed uploads. The list returned in the List Parts response is sorted by part range.

The List Parts operation supports pagination. By default, this operation returns up to 1,000 uploaded parts in the response. You should always check the response for a marker at which to continue the list; if there are no more items the marker is null . To return a list of parts that begins at a specific part, set the marker request parameter to the value you obtained from a previous List Parts request. You can also limit the number of parts returned in the response by specifying the limit parameter in the request.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identity and Access Management (IAM) .

For conceptual information and the underlying REST API, see Working with Archives in Amazon Glacier and List Parts in the Amazon Glacier Developer Guide .

This operation returns paginated results.

Creating a Request

listParts

Arguments

:: Text
lpAccountId

-> Text
lpVaultName

-> Text
lpUploadId

-> ListParts

Creates a value of ListParts with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lpMarker - An opaque string used for pagination. This value specifies the part at which the listing of parts should begin. Get the marker value from the response of a previous List Parts response. You need only include the marker if you are continuing the pagination of results started in a previous List Parts request.

lpLimit - The maximum number of parts to be returned. The default limit is 1000. The number of parts returned might be fewer than the specified limit, but the number of returned parts never exceeds the limit.

lpAccountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

lpVaultName - The name of the vault.

lpUploadId - The upload ID of the multipart upload.

data ListParts

Provides options for retrieving a list of parts of an archive that have been uploaded in a specific multipart upload.

See: listParts smart constructor.

Request Lenses

lpMarker :: Lens' ListParts (Maybe Text)

An opaque string used for pagination. This value specifies the part at which the listing of parts should begin. Get the marker value from the response of a previous List Parts response. You need only include the marker if you are continuing the pagination of results started in a previous List Parts request.

lpLimit :: Lens' ListParts (Maybe Text)

The maximum number of parts to be returned. The default limit is 1000. The number of parts returned might be fewer than the specified limit, but the number of returned parts never exceeds the limit.

lpAccountId :: Lens' ListParts Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

lpVaultName :: Lens' ListParts Text

The name of the vault.

lpUploadId :: Lens' ListParts Text

The upload ID of the multipart upload.

Destructuring the Response

listPartsResponse

Arguments

:: Int
lprsResponseStatus

-> ListPartsResponse

Creates a value of ListPartsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lprsParts - A list of the part sizes of the multipart upload. Each object in the array contains a RangeBytes and sha256-tree-hash name/value pair.

lprsMultipartUploadId - The ID of the upload to which the parts are associated.

lprsPartSizeInBytes - The part size in bytes. This is the same value that you specified in the Initiate Multipart Upload request.

lprsArchiveDescription - The description of the archive that was specified in the Initiate Multipart Upload request.

lprsVaultARN - The Amazon Resource Name (ARN) of the vault to which the multipart upload was initiated.

lprsMarker - An opaque string that represents where to continue pagination of the results. You use the marker in a new List Parts request to obtain more jobs in the list. If there are no more parts, this value is null .

lprsCreationDate - The UTC time at which the multipart upload was initiated.

lprsResponseStatus - -- | The response status code.

data ListPartsResponse

Contains the Amazon Glacier response to your request.

See: listPartsResponse smart constructor.

Response Lenses

lprsParts :: Lens' ListPartsResponse [PartListElement]

A list of the part sizes of the multipart upload. Each object in the array contains a RangeBytes and sha256-tree-hash name/value pair.

lprsMultipartUploadId :: Lens' ListPartsResponse (Maybe Text)

The ID of the upload to which the parts are associated.

lprsPartSizeInBytes :: Lens' ListPartsResponse (Maybe Integer)

The part size in bytes. This is the same value that you specified in the Initiate Multipart Upload request.

lprsArchiveDescription :: Lens' ListPartsResponse (Maybe Text)

The description of the archive that was specified in the Initiate Multipart Upload request.

lprsVaultARN :: Lens' ListPartsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the vault to which the multipart upload was initiated.

lprsMarker :: Lens' ListPartsResponse (Maybe Text)

An opaque string that represents where to continue pagination of the results. You use the marker in a new List Parts request to obtain more jobs in the list. If there are no more parts, this value is null .

lprsCreationDate :: Lens' ListPartsResponse (Maybe Text)

The UTC time at which the multipart upload was initiated.

lprsResponseStatus :: Lens' ListPartsResponse Int

- | The response status code.