Network.AWS.Glacier.ListVaults

Description

This operation lists all vaults owned by the calling user's account. The list returned in the response is ASCII-sorted by vault name.

By default, this operation returns up to 1,000 items. If there are more vaults to list, the response marker field contains the vault Amazon Resource Name (ARN) at which to continue the list with a new List Vaults request; otherwise, the marker field is null . To return a list of vaults that begins at a specific vault, set the marker request parameter to the vault ARN you obtained from a previous List Vaults request. You can also limit the number of vaults returned in the response by specifying the limit parameter in the request.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identity and Access Management (IAM) .

For conceptual information and underlying REST API, see Retrieving Vault Metadata in Amazon Glacier and List Vaults in the Amazon Glacier Developer Guide .

This operation returns paginated results.

Creating a Request

listVaults

Arguments

:: Text
lvAccountId

-> ListVaults

Creates a value of ListVaults with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lvMarker - A string used for pagination. The marker specifies the vault ARN after which the listing of vaults should begin.

lvLimit - The maximum number of vaults to be returned. The default limit is 1000. The number of vaults returned might be fewer than the specified limit, but the number of returned vaults never exceeds the limit.

lvAccountId - The AccountId value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens (-) in the ID.

data ListVaults

Provides options to retrieve the vault list owned by the calling user's account. The list provides metadata information for each vault.

See: listVaults smart constructor.

Request Lenses

lvMarker :: Lens' ListVaults (Maybe Text)

A string used for pagination. The marker specifies the vault ARN after which the listing of vaults should begin.

lvLimit :: Lens' ListVaults (Maybe Text)

The maximum number of vaults to be returned. The default limit is 1000. The number of vaults returned might be fewer than the specified limit, but the number of returned vaults never exceeds the limit.

lvAccountId :: Lens' ListVaults Text

The AccountId value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens (-) in the ID.

Destructuring the Response

listVaultsResponse

Arguments

:: Int
lvrsResponseStatus

-> ListVaultsResponse

Creates a value of ListVaultsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lvrsMarker - The vault ARN at which to continue pagination of the results. You use the marker in another List Vaults request to obtain more vaults in the list.

lvrsVaultList - List of vaults.

lvrsResponseStatus - -- | The response status code.

data ListVaultsResponse

Contains the Amazon Glacier response to your request.

See: listVaultsResponse smart constructor.

Response Lenses

lvrsMarker :: Lens' ListVaultsResponse (Maybe Text)

The vault ARN at which to continue pagination of the results. You use the marker in another List Vaults request to obtain more vaults in the list.

lvrsVaultList :: Lens' ListVaultsResponse [DescribeVaultOutput]

List of vaults.

lvrsResponseStatus :: Lens' ListVaultsResponse Int

- | The response status code.