Network.AWS.Glacier.SetVaultNotifications

Description

This operation configures notifications that will be sent when specific events happen to a vault. By default, you don't get any notifications.

To configure vault notifications, send a PUT request to the notification-configuration subresource of the vault. The request should include a JSON document that provides an Amazon SNS topic and specific events for which you want Amazon Glacier to send notifications to the topic.

Amazon SNS topics must grant permission to the vault to be allowed to publish notifications to the topic. You can configure a vault to publish a notification for the following vault events:

ArchiveRetrievalCompleted This event occurs when a job that was initiated for an archive retrieval is completed (InitiateJob ). The status of the completed job can be Succeeded or Failed. The notification sent to the SNS topic is the same output as returned from DescribeJob .

InventoryRetrievalCompleted This event occurs when a job that was initiated for an inventory retrieval is completed (InitiateJob ). The status of the completed job can be Succeeded or Failed. The notification sent to the SNS topic is the same output as returned from DescribeJob .

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identity and Access Management (IAM) .

For conceptual information and underlying REST API, see Configuring Vault Notifications in Amazon Glacier and Set Vault Notification Configuration in the Amazon Glacier Developer Guide .

Creating a Request

setVaultNotifications

Arguments

:: Text
svnAccountId

-> Text
svnVaultName

-> SetVaultNotifications

Creates a value of SetVaultNotifications with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

svnVaultNotificationConfig - Provides options for specifying notification configuration.

svnAccountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

svnVaultName - The name of the vault.

data SetVaultNotifications

Provides options to configure notifications that will be sent when specific events happen to a vault.

See: setVaultNotifications smart constructor.

Request Lenses

svnVaultNotificationConfig :: Lens' SetVaultNotifications (Maybe VaultNotificationConfig)

Provides options for specifying notification configuration.

svnAccountId :: Lens' SetVaultNotifications Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

svnVaultName :: Lens' SetVaultNotifications Text

The name of the vault.

Destructuring the Response

setVaultNotificationsResponse :: SetVaultNotificationsResponse

Creates a value of SetVaultNotificationsResponse with the minimum fields required to make a request.

data SetVaultNotificationsResponse

See: setVaultNotificationsResponse smart constructor.