Network.AWS.Glacier.Types

Description

Service Configuration

glacier :: Service

API version 2012-06-01 of the Amazon Glacier SDK configuration.

Errors

_PolicyEnforcedException :: AsError a => Getting (First ServiceError) a ServiceError

Returned if a retrieval job would exceed the current data policy's retrieval rate limit. For more information about data retrieval policies,

_InvalidParameterValueException :: AsError a => Getting (First ServiceError) a ServiceError

Returned if a parameter of the request is incorrectly specified.

_RequestTimeoutException :: AsError a => Getting (First ServiceError) a ServiceError

Returned if, when uploading an archive, Amazon Glacier times out while receiving the upload.

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the service cannot complete the request.

_InsufficientCapacityException :: AsError a => Getting (First ServiceError) a ServiceError

Returned if there is insufficient capacity to process this expedited request. This error only applies to expedited retrievals and not to standard or bulk retrievals.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the specified resource (such as a vault, upload ID, or job ID) doesn't exist.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the request results in a vault or account limit being exceeded.

_MissingParameterValueException :: AsError a => Getting (First ServiceError) a ServiceError

Returned if a required header or parameter is missing from the request.

ActionCode

data ActionCode

Constructors

ArchiveRetrieval

InventoryRetrieval

StatusCode

data StatusCode

Constructors

Failed

InProgress

Succeeded

ArchiveCreationOutput

data ArchiveCreationOutput

Contains the Amazon Glacier response to your request.

For information about the underlying REST API, see Upload Archive . For conceptual information, see Working with Archives in Amazon Glacier .

See: archiveCreationOutput smart constructor.

archiveCreationOutput :: ArchiveCreationOutput

Creates a value of ArchiveCreationOutput with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

acoArchiveId - The ID of the archive. This value is also included as part of the location.

acoChecksum - The checksum of the archive computed by Amazon Glacier.

acoLocation - The relative URI path of the newly added archive resource.

acoArchiveId :: Lens' ArchiveCreationOutput (Maybe Text)

The ID of the archive. This value is also included as part of the location.

acoChecksum :: Lens' ArchiveCreationOutput (Maybe Text)

The checksum of the archive computed by Amazon Glacier.

acoLocation :: Lens' ArchiveCreationOutput (Maybe Text)

The relative URI path of the newly added archive resource.

DataRetrievalPolicy

data DataRetrievalPolicy

Data retrieval policy.

See: dataRetrievalPolicy smart constructor.

dataRetrievalPolicy :: DataRetrievalPolicy

Creates a value of DataRetrievalPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drpRules - The policy rule. Although this is a list type, currently there must be only one rule, which contains a Strategy field and optionally a BytesPerHour field.

drpRules :: Lens' DataRetrievalPolicy [DataRetrievalRule]

The policy rule. Although this is a list type, currently there must be only one rule, which contains a Strategy field and optionally a BytesPerHour field.

DataRetrievalRule

data DataRetrievalRule

Data retrieval policy rule.

See: dataRetrievalRule smart constructor.

dataRetrievalRule :: DataRetrievalRule

Creates a value of DataRetrievalRule with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drrStrategy - The type of data retrieval policy to set. Valid values: BytesPerHour|FreeTier|None

drrBytesPerHour - The maximum number of bytes that can be retrieved in an hour. This field is required only if the value of the Strategy field is BytesPerHour . Your PUT operation will be rejected if the Strategy field is not set to BytesPerHour and you set this field.

drrStrategy :: Lens' DataRetrievalRule (Maybe Text)

The type of data retrieval policy to set. Valid values: BytesPerHour|FreeTier|None

drrBytesPerHour :: Lens' DataRetrievalRule (Maybe Integer)

The maximum number of bytes that can be retrieved in an hour. This field is required only if the value of the Strategy field is BytesPerHour . Your PUT operation will be rejected if the Strategy field is not set to BytesPerHour and you set this field.

DescribeVaultOutput

data DescribeVaultOutput

Contains the Amazon Glacier response to your request.

See: describeVaultOutput smart constructor.

describeVaultOutput :: DescribeVaultOutput

Creates a value of DescribeVaultOutput with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvoVaultName - The name of the vault.

dvoSizeInBytes - Total size, in bytes, of the archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.

dvoLastInventoryDate - The Universal Coordinated Time (UTC) date when Amazon Glacier completed the last vault inventory. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z .

dvoVaultARN - The Amazon Resource Name (ARN) of the vault.

dvoCreationDate - The Universal Coordinated Time (UTC) date when the vault was created. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z .

dvoNumberOfArchives - The number of archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.

dvoVaultName :: Lens' DescribeVaultOutput (Maybe Text)

The name of the vault.

dvoSizeInBytes :: Lens' DescribeVaultOutput (Maybe Integer)

Total size, in bytes, of the archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.

dvoLastInventoryDate :: Lens' DescribeVaultOutput (Maybe Text)

The Universal Coordinated Time (UTC) date when Amazon Glacier completed the last vault inventory. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z .

dvoVaultARN :: Lens' DescribeVaultOutput (Maybe Text)

The Amazon Resource Name (ARN) of the vault.

dvoCreationDate :: Lens' DescribeVaultOutput (Maybe Text)

The Universal Coordinated Time (UTC) date when the vault was created. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z .

dvoNumberOfArchives :: Lens' DescribeVaultOutput (Maybe Integer)

The number of archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.

GlacierJobDescription

data GlacierJobDescription

Describes an Amazon Glacier job.

See: glacierJobDescription smart constructor.

glacierJobDescription :: GlacierJobDescription

Creates a value of GlacierJobDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gjdSHA256TreeHash - For an ArchiveRetrieval job, it is the checksum of the archive. Otherwise, the value is null. The SHA256 tree hash value for the requested range of an archive. If the Initiate a Job request for an archive specified a tree-hash aligned range, then this field returns a value. For the specific case when the whole archive is retrieved, this value is the same as the ArchiveSHA256TreeHash value. This field is null in the following situations: * Archive retrieval jobs that specify a range that is not tree-hash aligned. * Archival jobs that specify a range that is equal to the whole archive and the job status is InProgress. * Inventory jobs.

gjdArchiveId - For an ArchiveRetrieval job, this is the archive ID requested for download. Otherwise, this field is null.

gjdJobId - An opaque string that identifies an Amazon Glacier job.

gjdRetrievalByteRange - The retrieved byte range for archive retrieval jobs in the form "StartByteValue -EndByteValue " If no range was specified in the archive retrieval, then the whole archive is retrieved and StartByteValue equals 0 and EndByteValue equals the size of the archive minus 1. For inventory retrieval jobs this field is null.

gjdInventoryRetrievalParameters - Parameters used for range inventory retrieval.

gjdAction - The job type. It is either ArchiveRetrieval or InventoryRetrieval.

gjdJobDescription - The job description you provided when you initiated the job.

gjdSNSTopic - An Amazon Simple Notification Service (Amazon SNS) topic that receives notification.

gjdStatusMessage - A friendly message that describes the job status.

gjdVaultARN - The Amazon Resource Name (ARN) of the vault from which the archive retrieval was requested.

gjdTier - The retrieval option to use for the archive retrieval. Valid values are Expedited , Standard , or Bulk . Standard is the default.

gjdArchiveSHA256TreeHash - The SHA256 tree hash of the entire archive for an archive retrieval. For inventory retrieval jobs, this field is null.

gjdCreationDate - The UTC date when the job was created. A string representation of ISO 8601 date format, for example, "2012-03-20T17:03:43.221Z".

gjdCompleted - The job status. When a job is completed, you get the job's output.

gjdCompletionDate - The UTC time that the archive retrieval request completed. While the job is in progress, the value will be null.

gjdInventorySizeInBytes - For an InventoryRetrieval job, this is the size in bytes of the inventory requested for download. For the ArchiveRetrieval job, the value is null.

gjdArchiveSizeInBytes - For an ArchiveRetrieval job, this is the size in bytes of the archive being requested for download. For the InventoryRetrieval job, the value is null.

gjdStatusCode - The status code can be InProgress, Succeeded, or Failed, and indicates the status of the job.

gjdSHA256TreeHash :: Lens' GlacierJobDescription (Maybe Text)

For an ArchiveRetrieval job, it is the checksum of the archive. Otherwise, the value is null. The SHA256 tree hash value for the requested range of an archive. If the Initiate a Job request for an archive specified a tree-hash aligned range, then this field returns a value. For the specific case when the whole archive is retrieved, this value is the same as the ArchiveSHA256TreeHash value. This field is null in the following situations: * Archive retrieval jobs that specify a range that is not tree-hash aligned. * Archival jobs that specify a range that is equal to the whole archive and the job status is InProgress. * Inventory jobs.

gjdArchiveId :: Lens' GlacierJobDescription (Maybe Text)

For an ArchiveRetrieval job, this is the archive ID requested for download. Otherwise, this field is null.

gjdJobId :: Lens' GlacierJobDescription (Maybe Text)

An opaque string that identifies an Amazon Glacier job.

gjdRetrievalByteRange :: Lens' GlacierJobDescription (Maybe Text)

The retrieved byte range for archive retrieval jobs in the form "StartByteValue -EndByteValue " If no range was specified in the archive retrieval, then the whole archive is retrieved and StartByteValue equals 0 and EndByteValue equals the size of the archive minus 1. For inventory retrieval jobs this field is null.

gjdInventoryRetrievalParameters :: Lens' GlacierJobDescription (Maybe InventoryRetrievalJobDescription)

Parameters used for range inventory retrieval.

gjdAction :: Lens' GlacierJobDescription (Maybe ActionCode)

The job type. It is either ArchiveRetrieval or InventoryRetrieval.

gjdJobDescription :: Lens' GlacierJobDescription (Maybe Text)

The job description you provided when you initiated the job.

gjdSNSTopic :: Lens' GlacierJobDescription (Maybe Text)

An Amazon Simple Notification Service (Amazon SNS) topic that receives notification.

gjdStatusMessage :: Lens' GlacierJobDescription (Maybe Text)

A friendly message that describes the job status.

gjdVaultARN :: Lens' GlacierJobDescription (Maybe Text)

The Amazon Resource Name (ARN) of the vault from which the archive retrieval was requested.

gjdTier :: Lens' GlacierJobDescription (Maybe Text)

The retrieval option to use for the archive retrieval. Valid values are Expedited , Standard , or Bulk . Standard is the default.

gjdArchiveSHA256TreeHash :: Lens' GlacierJobDescription (Maybe Text)

The SHA256 tree hash of the entire archive for an archive retrieval. For inventory retrieval jobs, this field is null.

gjdCreationDate :: Lens' GlacierJobDescription (Maybe Text)

The UTC date when the job was created. A string representation of ISO 8601 date format, for example, "2012-03-20T17:03:43.221Z".

gjdCompleted :: Lens' GlacierJobDescription (Maybe Bool)

The job status. When a job is completed, you get the job's output.

gjdCompletionDate :: Lens' GlacierJobDescription (Maybe Text)

The UTC time that the archive retrieval request completed. While the job is in progress, the value will be null.

gjdInventorySizeInBytes :: Lens' GlacierJobDescription (Maybe Integer)

For an InventoryRetrieval job, this is the size in bytes of the inventory requested for download. For the ArchiveRetrieval job, the value is null.

gjdArchiveSizeInBytes :: Lens' GlacierJobDescription (Maybe Integer)

For an ArchiveRetrieval job, this is the size in bytes of the archive being requested for download. For the InventoryRetrieval job, the value is null.

gjdStatusCode :: Lens' GlacierJobDescription (Maybe StatusCode)

The status code can be InProgress, Succeeded, or Failed, and indicates the status of the job.

InventoryRetrievalJobDescription

data InventoryRetrievalJobDescription

Describes the options for a range inventory retrieval job.

See: inventoryRetrievalJobDescription smart constructor.

inventoryRetrievalJobDescription :: InventoryRetrievalJobDescription

Creates a value of InventoryRetrievalJobDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

irjdFormat - The output format for the vault inventory list, which is set by the InitiateJob request when initiating a job to retrieve a vault inventory. Valid values are CSV and JSON .

irjdEndDate - The end of the date range in UTC for vault inventory retrieval that includes archives created before this date. This value should be a string in the ISO 8601 date format, for example 2013-03-20T17:03:43Z .

irjdStartDate - The start of the date range in Universal Coordinated Time (UTC) for vault inventory retrieval that includes archives created on or after this date. This value should be a string in the ISO 8601 date format, for example 2013-03-20T17:03:43Z .

irjdMarker - An opaque string that represents where to continue pagination of the vault inventory retrieval results. You use the marker in a new InitiateJob request to obtain additional inventory items. If there are no more inventory items, this value is null . For more information, see Range Inventory Retrieval .

irjdLimit - The maximum number of inventory items returned per vault inventory retrieval request. This limit is set when initiating the job with the a InitiateJob request.

irjdFormat :: Lens' InventoryRetrievalJobDescription (Maybe Text)

The output format for the vault inventory list, which is set by the InitiateJob request when initiating a job to retrieve a vault inventory. Valid values are CSV and JSON .

irjdEndDate :: Lens' InventoryRetrievalJobDescription (Maybe Text)

The end of the date range in UTC for vault inventory retrieval that includes archives created before this date. This value should be a string in the ISO 8601 date format, for example 2013-03-20T17:03:43Z .

irjdStartDate :: Lens' InventoryRetrievalJobDescription (Maybe Text)

The start of the date range in Universal Coordinated Time (UTC) for vault inventory retrieval that includes archives created on or after this date. This value should be a string in the ISO 8601 date format, for example 2013-03-20T17:03:43Z .

irjdMarker :: Lens' InventoryRetrievalJobDescription (Maybe Text)

An opaque string that represents where to continue pagination of the vault inventory retrieval results. You use the marker in a new InitiateJob request to obtain additional inventory items. If there are no more inventory items, this value is null . For more information, see Range Inventory Retrieval .

irjdLimit :: Lens' InventoryRetrievalJobDescription (Maybe Text)

The maximum number of inventory items returned per vault inventory retrieval request. This limit is set when initiating the job with the a InitiateJob request.

InventoryRetrievalJobInput

data InventoryRetrievalJobInput

Provides options for specifying a range inventory retrieval job.

See: inventoryRetrievalJobInput smart constructor.

inventoryRetrievalJobInput :: InventoryRetrievalJobInput

Creates a value of InventoryRetrievalJobInput with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

irjiEndDate - The end of the date range in UTC for vault inventory retrieval that includes archives created before this date. This value should be a string in the ISO 8601 date format, for example 2013-03-20T17:03:43Z .

irjiStartDate - The start of the date range in UTC for vault inventory retrieval that includes archives created on or after this date. This value should be a string in the ISO 8601 date format, for example 2013-03-20T17:03:43Z .

irjiMarker - An opaque string that represents where to continue pagination of the vault inventory retrieval results. You use the marker in a new InitiateJob request to obtain additional inventory items. If there are no more inventory items, this value is null .

irjiLimit - Specifies the maximum number of inventory items returned per vault inventory retrieval request. Valid values are greater than or equal to 1.

irjiEndDate :: Lens' InventoryRetrievalJobInput (Maybe Text)

The end of the date range in UTC for vault inventory retrieval that includes archives created before this date. This value should be a string in the ISO 8601 date format, for example 2013-03-20T17:03:43Z .

irjiStartDate :: Lens' InventoryRetrievalJobInput (Maybe Text)

The start of the date range in UTC for vault inventory retrieval that includes archives created on or after this date. This value should be a string in the ISO 8601 date format, for example 2013-03-20T17:03:43Z .

irjiMarker :: Lens' InventoryRetrievalJobInput (Maybe Text)

An opaque string that represents where to continue pagination of the vault inventory retrieval results. You use the marker in a new InitiateJob request to obtain additional inventory items. If there are no more inventory items, this value is null .

irjiLimit :: Lens' InventoryRetrievalJobInput (Maybe Text)

Specifies the maximum number of inventory items returned per vault inventory retrieval request. Valid values are greater than or equal to 1.

JobParameters

data JobParameters

Provides options for defining a job.

See: jobParameters smart constructor.

jobParameters :: JobParameters

Creates a value of JobParameters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

jpArchiveId - The ID of the archive that you want to retrieve. This field is required only if Type is set to archive-retrieval. An error occurs if you specify this request parameter for an inventory retrieval job request.

jpFormat - When initiating a job to retrieve a vault inventory, you can optionally add this parameter to your request to specify the output format. If you are initiating an inventory job and do not specify a Format field, JSON is the default format. Valid values are CSV and JSON.

jpRetrievalByteRange - The byte range to retrieve for an archive retrieval. in the form "StartByteValue -EndByteValue " If not specified, the whole archive is retrieved. If specified, the byte range must be megabyte (1024*1024) aligned which means that StartByteValue must be divisible by 1 MB and EndByteValue plus 1 must be divisible by 1 MB or be the end of the archive specified as the archive byte size value minus 1. If RetrievalByteRange is not megabyte aligned, this operation returns a 400 response. An error occurs if you specify this field for an inventory retrieval job request.

jpInventoryRetrievalParameters - Input parameters used for range inventory retrieval.

jpSNSTopic - The Amazon SNS topic ARN to which Amazon Glacier sends a notification when the job is completed and the output is ready for you to download. The specified topic publishes the notification to its subscribers. The SNS topic must exist.

jpTier - The retrieval option to use for the archive retrieval. Valid values are Expedited , Standard , or Bulk . Standard is the default.

jpType - The job type. You can initiate a job to retrieve an archive or get an inventory of a vault. Valid values are "archive-retrieval" and "inventory-retrieval".

jpDescription - The optional description for the job. The description must be less than or equal to 1,024 bytes. The allowable characters are 7-bit ASCII without control codes-specifically, ASCII values 32-126 decimal or 0x20-0x7E hexadecimal.

jpArchiveId :: Lens' JobParameters (Maybe Text)

The ID of the archive that you want to retrieve. This field is required only if Type is set to archive-retrieval. An error occurs if you specify this request parameter for an inventory retrieval job request.

jpFormat :: Lens' JobParameters (Maybe Text)

When initiating a job to retrieve a vault inventory, you can optionally add this parameter to your request to specify the output format. If you are initiating an inventory job and do not specify a Format field, JSON is the default format. Valid values are CSV and JSON.

jpRetrievalByteRange :: Lens' JobParameters (Maybe Text)

The byte range to retrieve for an archive retrieval. in the form "StartByteValue -EndByteValue " If not specified, the whole archive is retrieved. If specified, the byte range must be megabyte (1024*1024) aligned which means that StartByteValue must be divisible by 1 MB and EndByteValue plus 1 must be divisible by 1 MB or be the end of the archive specified as the archive byte size value minus 1. If RetrievalByteRange is not megabyte aligned, this operation returns a 400 response. An error occurs if you specify this field for an inventory retrieval job request.

jpInventoryRetrievalParameters :: Lens' JobParameters (Maybe InventoryRetrievalJobInput)

Input parameters used for range inventory retrieval.

jpSNSTopic :: Lens' JobParameters (Maybe Text)

The Amazon SNS topic ARN to which Amazon Glacier sends a notification when the job is completed and the output is ready for you to download. The specified topic publishes the notification to its subscribers. The SNS topic must exist.

jpTier :: Lens' JobParameters (Maybe Text)

The retrieval option to use for the archive retrieval. Valid values are Expedited , Standard , or Bulk . Standard is the default.

jpType :: Lens' JobParameters (Maybe Text)

The job type. You can initiate a job to retrieve an archive or get an inventory of a vault. Valid values are "archive-retrieval" and "inventory-retrieval".

jpDescription :: Lens' JobParameters (Maybe Text)

The optional description for the job. The description must be less than or equal to 1,024 bytes. The allowable characters are 7-bit ASCII without control codes-specifically, ASCII values 32-126 decimal or 0x20-0x7E hexadecimal.

PartListElement

data PartListElement

A list of the part sizes of the multipart upload.

See: partListElement smart constructor.

partListElement :: PartListElement

Creates a value of PartListElement with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pleSHA256TreeHash - The SHA256 tree hash value that Amazon Glacier calculated for the part. This field is never null .

pleRangeInBytes - The byte range of a part, inclusive of the upper value of the range.

pleSHA256TreeHash :: Lens' PartListElement (Maybe Text)

The SHA256 tree hash value that Amazon Glacier calculated for the part. This field is never null .

pleRangeInBytes :: Lens' PartListElement (Maybe Text)

The byte range of a part, inclusive of the upper value of the range.

ProvisionedCapacityDescription

data ProvisionedCapacityDescription

The definition for a provisioned capacity unit.

See: provisionedCapacityDescription smart constructor.

provisionedCapacityDescription :: ProvisionedCapacityDescription

Creates a value of ProvisionedCapacityDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pcdCapacityId - The ID that identifies the provisioned capacity unit.

pcdStartDate - The date that the provisioned capacity unit was purchased, in Universal Coordinated Time (UTC).

pcdExpirationDate - The date that the provisioned capacity unit expires, in Universal Coordinated Time (UTC).

pcdCapacityId :: Lens' ProvisionedCapacityDescription (Maybe Text)

The ID that identifies the provisioned capacity unit.

pcdStartDate :: Lens' ProvisionedCapacityDescription (Maybe Text)

The date that the provisioned capacity unit was purchased, in Universal Coordinated Time (UTC).

pcdExpirationDate :: Lens' ProvisionedCapacityDescription (Maybe Text)

The date that the provisioned capacity unit expires, in Universal Coordinated Time (UTC).

UploadListElement

data UploadListElement

A list of in-progress multipart uploads for a vault.

See: uploadListElement smart constructor.

uploadListElement :: UploadListElement

Creates a value of UploadListElement with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uleMultipartUploadId - The ID of a multipart upload.

ulePartSizeInBytes - The part size, in bytes, specified in the Initiate Multipart Upload request. This is the size of all the parts in the upload except the last part, which may be smaller than this size.

uleArchiveDescription - The description of the archive that was specified in the Initiate Multipart Upload request.

uleVaultARN - The Amazon Resource Name (ARN) of the vault that contains the archive.

uleCreationDate - The UTC time at which the multipart upload was initiated.

uleMultipartUploadId :: Lens' UploadListElement (Maybe Text)

The ID of a multipart upload.

ulePartSizeInBytes :: Lens' UploadListElement (Maybe Integer)

The part size, in bytes, specified in the Initiate Multipart Upload request. This is the size of all the parts in the upload except the last part, which may be smaller than this size.

uleArchiveDescription :: Lens' UploadListElement (Maybe Text)

The description of the archive that was specified in the Initiate Multipart Upload request.

uleVaultARN :: Lens' UploadListElement (Maybe Text)

The Amazon Resource Name (ARN) of the vault that contains the archive.

uleCreationDate :: Lens' UploadListElement (Maybe Text)

The UTC time at which the multipart upload was initiated.

VaultAccessPolicy

data VaultAccessPolicy

Contains the vault access policy.

See: vaultAccessPolicy smart constructor.

vaultAccessPolicy :: VaultAccessPolicy

Creates a value of VaultAccessPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vapPolicy - The vault access policy.

vapPolicy :: Lens' VaultAccessPolicy (Maybe Text)

The vault access policy.

VaultLockPolicy

data VaultLockPolicy

Contains the vault lock policy.

See: vaultLockPolicy smart constructor.

vaultLockPolicy :: VaultLockPolicy

Creates a value of VaultLockPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vlpPolicy - The vault lock policy.

vlpPolicy :: Lens' VaultLockPolicy (Maybe Text)

The vault lock policy.

VaultNotificationConfig

data VaultNotificationConfig

Represents a vault's notification configuration.

See: vaultNotificationConfig smart constructor.

vaultNotificationConfig :: VaultNotificationConfig

Creates a value of VaultNotificationConfig with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vncSNSTopic - The Amazon Simple Notification Service (Amazon SNS) topic Amazon Resource Name (ARN).

vncEvents - A list of one or more events for which Amazon Glacier will send a notification to the specified Amazon SNS topic.

vncSNSTopic :: Lens' VaultNotificationConfig (Maybe Text)

The Amazon Simple Notification Service (Amazon SNS) topic Amazon Resource Name (ARN).

vncEvents :: Lens' VaultNotificationConfig [Text]

A list of one or more events for which Amazon Glacier will send a notification to the specified Amazon SNS topic.