Network.AWS.Glacier.UploadArchive

Description

This operation adds an archive to a vault. This is a synchronous operation, and for a successful upload, your data is durably persisted. Amazon Glacier returns the archive ID in the x-amz-archive-id header of the response.

You must use the archive ID to access your data in Amazon Glacier. After you upload an archive, you should save the archive ID returned so that you can retrieve or delete the archive later. Besides saving the archive ID, you can also index it and give it a friendly name to allow for better searching. You can also use the optional archive description field to specify how the archive is referred to in an external index of archives, such as you might create in Amazon DynamoDB. You can also get the vault inventory to obtain a list of archive IDs in a vault. For more information, see InitiateJob .

You must provide a SHA256 tree hash of the data you are uploading. For information about computing a SHA256 tree hash, see Computing Checksums .

You can optionally specify an archive description of up to 1,024 printable ASCII characters. You can get the archive description when you either retrieve the archive or get the vault inventory. For more information, see InitiateJob . Amazon Glacier does not interpret the description in any way. An archive description does not need to be unique. You cannot use the description to retrieve or sort the archive list.

Archives are immutable. After you upload an archive, you cannot edit the archive or its description.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identity and Access Management (IAM) .

For conceptual information and underlying REST API, see Uploading an Archive in Amazon Glacier and Upload Archive in the Amazon Glacier Developer Guide .

Creating a Request

uploadArchive

Arguments

:: Text
uaVaultName

-> Text
uaAccountId

-> HashedBody
uaBody

-> UploadArchive

Creates a value of UploadArchive with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uaChecksum - The SHA256 tree hash of the data being uploaded.

uaArchiveDescription - The optional description of the archive you are uploading.

uaVaultName - The name of the vault.

uaAccountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

uaBody - The data to upload.

data UploadArchive

Provides options to add an archive to a vault.

See: uploadArchive smart constructor.

Request Lenses

uaChecksum :: Lens' UploadArchive (Maybe Text)

The SHA256 tree hash of the data being uploaded.

uaArchiveDescription :: Lens' UploadArchive (Maybe Text)

The optional description of the archive you are uploading.

uaVaultName :: Lens' UploadArchive Text

The name of the vault.

uaAccountId :: Lens' UploadArchive Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

uaBody :: Lens' UploadArchive HashedBody

The data to upload.

Destructuring the Response

archiveCreationOutput :: ArchiveCreationOutput

Creates a value of ArchiveCreationOutput with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

acoArchiveId - The ID of the archive. This value is also included as part of the location.

acoChecksum - The checksum of the archive computed by Amazon Glacier.

acoLocation - The relative URI path of the newly added archive resource.

data ArchiveCreationOutput

Contains the Amazon Glacier response to your request.

For information about the underlying REST API, see Upload Archive . For conceptual information, see Working with Archives in Amazon Glacier .

See: archiveCreationOutput smart constructor.

Response Lenses

acoArchiveId :: Lens' ArchiveCreationOutput (Maybe Text)

The ID of the archive. This value is also included as part of the location.

acoChecksum :: Lens' ArchiveCreationOutput (Maybe Text)

The checksum of the archive computed by Amazon Glacier.

acoLocation :: Lens' ArchiveCreationOutput (Maybe Text)

The relative URI path of the newly added archive resource.