Network.AWS.Glacier.UploadMultipartPart

Description

This operation uploads a part of an archive. You can upload archive parts in any order. You can also upload them in parallel. You can upload up to 10,000 parts for a multipart upload.

Amazon Glacier rejects your upload part request if any of the following conditions is true:

SHA256 tree hash does not match To ensure that part data is not corrupted in transmission, you compute a SHA256 tree hash of the part and include it in your request. Upon receiving the part data, Amazon Glacier also computes a SHA256 tree hash. If these hash values don't match, the operation fails. For information about computing a SHA256 tree hash, see Computing Checksums .

Part size does not match The size of each part except the last must match the size specified in the corresponding InitiateMultipartUpload request. The size of the last part must be the same size as, or smaller than, the specified size.

Range does not align The byte range value in the request does not align with the part size specified in the corresponding initiate request. For example, if you specify a part size of 4194304 bytes (4 MB), then 0 to 4194303 bytes (4 MB - 1) and 4194304 (4 MB) to 8388607 (8 MB - 1) are valid part ranges. However, if you set a range value of 2 MB to 6 MB, the range does not align with the part size and the upload will fail.

This operation is idempotent. If you upload the same part multiple times, the data included in the most recent request overwrites the previously uploaded data.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identity and Access Management (IAM) .

For conceptual information and underlying REST API, see Uploading Large Archives in Parts (Multipart Upload) and Upload Part in the Amazon Glacier Developer Guide .

Creating a Request

uploadMultipartPart

Arguments

:: Text
umpAccountId

-> Text
umpVaultName

-> Text
umpUploadId

-> HashedBody
umpBody

-> UploadMultipartPart

Creates a value of UploadMultipartPart with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

umpChecksum - The SHA256 tree hash of the data being uploaded.

umpRange - Identifies the range of bytes in the assembled archive that will be uploaded in this part. Amazon Glacier uses this information to assemble the archive in the proper sequence. The format of this header follows RFC 2616. An example header is Content-Range:bytes 0-4194303/*.

umpAccountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

umpVaultName - The name of the vault.

umpUploadId - The upload ID of the multipart upload.

umpBody - The data to upload.

data UploadMultipartPart

Provides options to upload a part of an archive in a multipart upload operation.

See: uploadMultipartPart smart constructor.

Request Lenses

umpChecksum :: Lens' UploadMultipartPart (Maybe Text)

The SHA256 tree hash of the data being uploaded.

umpRange :: Lens' UploadMultipartPart (Maybe Text)

Identifies the range of bytes in the assembled archive that will be uploaded in this part. Amazon Glacier uses this information to assemble the archive in the proper sequence. The format of this header follows RFC 2616. An example header is Content-Range:bytes 0-4194303/*.

umpAccountId :: Lens' UploadMultipartPart Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '- ' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens (-) in the ID.

umpVaultName :: Lens' UploadMultipartPart Text

The name of the vault.

umpUploadId :: Lens' UploadMultipartPart Text

The upload ID of the multipart upload.

umpBody :: Lens' UploadMultipartPart HashedBody

The data to upload.

Destructuring the Response

uploadMultipartPartResponse

Arguments

:: Int
umprsResponseStatus

-> UploadMultipartPartResponse

Creates a value of UploadMultipartPartResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

umprsChecksum - The SHA256 tree hash that Amazon Glacier computed for the uploaded part.

umprsResponseStatus - -- | The response status code.

data UploadMultipartPartResponse

Contains the Amazon Glacier response to your request.

See: uploadMultipartPartResponse smart constructor.

Response Lenses

umprsChecksum :: Lens' UploadMultipartPartResponse (Maybe Text)

The SHA256 tree hash that Amazon Glacier computed for the uploaded part.

umprsResponseStatus :: Lens' UploadMultipartPartResponse Int

- | The response status code.